==> R/2_phyloseq_object.Rmd <==
---
title: "2_phyloseq_object"
output: html_document
---

```{r phyloseq}
library(phyloseq); packageVersion("phyloseq")
library(ggplot2); packageVersion("ggplot2")
library(Biostrings)
library(readr)
library(readxl)
library(readr)

otu_table <- read_delim("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/otu_table.csv", 
    delim = ";", escape_double = FALSE, trim_ws = TRUE)
tax_table <- read_delim("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/tax_table.csv", 
    delim = ";", escape_double = FALSE, trim_ws = TRUE)

metadata <- read_excel("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/metadata.xlsx")
tree<- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/phylo_tree.rds")
dna<- readDNAStringSet("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ASV.fasta")


otu_table <- as.data.frame(otu_table)
rownames(otu_table) <- otu_table$...1
otu_ok <- subset(otu_table, select = -c(CV1_S1, CV16_S2, CR23_S3, CR24_S4, CC34_36_S5, CC37_1_S6, CCR55_S7, CCR56_S8, ...1))
otu_ok <- otu_ok[-1827, ]

tax_table <- as.data.frame(tax_table)
rownames(tax_table) <- tax_table$...1
tax_table$Species<- ifelse(is.na(tax_table$Species...8) == TRUE, tax_table$Species...9, ifelse(is.na(tax_table$Species...8) == FALSE, tax_table$Species...8, "NA"))
tax_table_ok <- subset(tax_table, select = -c(Species...8, Species...9, ...1))

 OTU = otu_table(otu_ok, taxa_are_rows = TRUE)
 TAX = tax_table(as.matrix(tax_table_ok))
 SAMPLE = sample_data(metadata)
 TREE = tree<- tree$tree
 
rownames(SAMPLE) <-metadata$SampleID

 ps_1 <- phyloseq(OTU, TAX, SAMPLE, TREE)
 ps_1
saveRDS(ps_1, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_1.Rds")  
saveRDS(otu_ok, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/otu_ok.Rds")  
saveRDS(tax_table_ok, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/tax_table_ok.Rds") 

#Rarefaction curve
S <- specnumber(t(otu_ok)) # observed number of species
(raremax <- min(rowSums(t(otu_ok))))
Srare <- rarefy(t(otu_ok), raremax)
plot(S, Srare, xlab = "Observed No. of Species", ylab = "Rarefied No. of Species")
abline(0, 1)
rarecurve(t(otu_ok), step = 20, sample = raremax, col = "black", cex = 0.6)
```
