==> R/3_Preprocessing_and_AlphaDiv.Rmd <==
---
title: "3_AlphaDiversity_and_DataTransformations"
output: html_document
date: "2023-04-26"
---
```{r Libraries}
library(phyloseq)
library(microbiome)
library(tidyverse)
library(ggpubr)
library(rstatix)
library(data.table)
library(knitr)
library(grid)
library(gridExtra)
```

#Saving processed phyloseq objects
```{r rarefaction}
ps_1 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_1.Rds")
otu_ok <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/otu_ok.Rds")
# keep result reproductive
 set.seed(111)
 ps.rarefied = rarefy_even_depth(ps_1, rngseed=1, replace=F)
 ps.rarefied

```

```{r Counts}
reads_per_sample <- as.matrix(rowSums(t(otu_ok)))
reads_per_sample_after_raref <- as.matrix(rowSums(t(as.matrix(otu_table(ps.rarefied)))))
```

```{r Clean}
 # Remove samples with less than MINREADS from phyloseq object
ps_2 <- prune_taxa(taxa_sums(ps_1) > 0, ps_1)
# merge all taxa that are detected rare
total_samples <- phyloseq::nsamples(ps_2)
ps_2
saveRDS(ps_2, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_2.Rds")

 # Remove samples with less than MINREADS from phyloseq object
ps.rarefied.clean <- prune_taxa(taxa_sums(ps.rarefied) > 1, ps.rarefied)
# merge all taxa that are detected rare
total_samples2 <- phyloseq::nsamples(ps.rarefied.clean)
ps.rarefied.clean

ps_2_genus <- tax_glom(ps_2, taxrank = 'Genus', NArm = FALSE)
ps_raref_genus <- tax_glom(ps.rarefied.clean, taxrank = 'Genus', NArm = FALSE)
```

#Trees
```{r random tress}
myTaxa = names(sort(taxa_sums(ps_2), decreasing = TRUE)[1:100])
ex1 = prune_taxa(myTaxa, ps_2)
plot(phy_tree(ex1), show.node.label = FALSE)
plot_tree(ex1, color = "Group", label.tips = "Phylum", ladderize = "left", justify = "left" , size = "Abundance")
```


#Number of taxa per group
```{r descriptive analysis}
ps_1 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_1.Rds")
ps_1
ps_2 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_2.Rds")
ps_2
ps.ctr<-ps_2 %>% subset_samples(Group %in% c("ctr"))
ps.ctr<-prune_taxa(taxa_sums(ps.ctr) > 0, ps.ctr)
ps.ctr
ps.ctrrif<-ps_2 %>% subset_samples(Group %in% c("ctr+rif"))
ps.ctrrif<-prune_taxa(taxa_sums(ps.ctrrif) > 0, ps.ctrrif)
ps.ctrrif
ps.ccl4<-ps_2 %>% subset_samples(Group %in% c("ccl4"))
ps.ccl4<-prune_taxa(taxa_sums(ps.ccl4) > 0, ps.ccl4)
ps.ccl4
ps.ccl4rif<-ps_2 %>% subset_samples(Group %in% c("ccl4+rif"))
ps.ccl4rif<-prune_taxa(taxa_sums(ps.ccl4rif) > 0, ps.ccl4rif)
ps.ccl4rif
```

#Dominant taxa
```{r descriptive analysis 2}
library(repmod)
#Phylum level
ps_2 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_2.Rds")
ps.rel<-ps_2 %>% microbiome::transform("compositional") %>% tax_glom(taxrank = 'Phylum', NArm = FALSE) 
# generate a vector containing the full taxonomy path for all OTUs
wholetax <- do.call(paste, c(as.data.frame(tax_table(ps.rel))
                             [c("Kingdom", "Phylum")], 
                             sep = "__"))  # to distinguish from "_" within tax ranks

# turn the otu_table into a data.frame
otu_export <- as.data.frame(t(otu_table(ps.rel)))
tmp <- names(otu_export)

# paste wholetax and OTU_ids together
for(i in 1:length(tmp)){
    names(tmp)[i] = paste(wholetax[i], tmp[i], sep = "__")
}
# overwrite old names
names(otu_export) <- names(tmp)
otu_export<-t(otu_export)
View(as.data.frame(rowSums(otu_export)))
otu_export1<-as.data.frame(rowSums(otu_export))
make_csv_table(otu_export1, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/Dominant_phyla", info = "Dominant_phyla")

#Family level
ps_2 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_2.Rds")
ps.rel2<-ps_2 %>% microbiome::transform("compositional") %>% tax_glom(taxrank = 'Family', NArm = FALSE)
# generate a vector containing the full taxonomy path for all OTUs
wholetax <- do.call(paste, c(as.data.frame(tax_table(ps.rel2))
                             [c("Kingdom", "Phylum", "Class", "Order", "Family")], 
                             sep = "__"))  # to distinguish from "_" within tax ranks

# turn the otu_table into a data.frame
otu_export.fam <- as.data.frame(t(otu_table(ps.rel2)))
tmp <- names(otu_export.fam)

# paste wholetax and OTU_ids together
for(i in 1:length(tmp)){
    names(tmp)[i] = paste(wholetax[i], tmp[i], sep = "__")
}
# overwrite old names
names(otu_export.fam) <- names(tmp)
otu_export.fam<-t(otu_export.fam)
View(as.data.frame(rowSums(otu_export.fam)))
otu_export2<-as.data.frame(rowSums(otu_export.fam))
make_csv_table(otu_export2, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/Dominant_families", info = "Dominant_fam")
```




It is important to not use filtered data because many richness estimates are modeled on singletons and doubletons in the occurrence table. So, you need to leave them in the dataset if you want a meaningful estimate.
Moreover, we usually not using normalized data because we want to assess the diversity on the raw data and we are not comparing samples to each other but only assessing diversity within each sample.

I have tested with and without rarefaction and filtering the zeros and without and the differences stay consisent.

#Alpha Diversity 
###Alpha Diversity statistical tests (check)
```{r Alpha Diversity statistical tests}
ps_2 <- readRDS("~/Library/CloudStorage/OneDrive-UniversitatdeValencia/Doctorado/Estudios/CIPF_Estudio1_2018/Rats_HE_CCL4model/output/ps_2.Rds")
require(lme4)
require("lmtest")
library(tidyverse)
Alpha<-estimate_richness(ps_2, measures = c("Observed", "Chao1","Shannon", "Fisher"))

Alpha%>%
  rownames_to_column()%>%
  dplyr::rename(SampleID = rowname)->tmp1

as_tibble(sample_data(ps_2))->tmp2

tmp1<-inner_join(tmp1, tmp2, by="SampleID")
rownames(tmp1)<- tmp1$SampleID
tmp1$SampleID<- NULL
sdt<- tmp1
rm(tmp1,tmp2)

#Differences between groups (general): Chao1
full<-lme4::lmer(Chao1 ~ Group + (1|Batch), sdt)
null<-lme4::lmer(Chao1 ~ (1|Batch), sdt)
lrtest(null, full)
#Model 1: Chao1 ~ (1 | Batch)
#Model 2: Chao1 ~ Group + (1 | Batch)
  #Df  LogLik Df  Chisq Pr(>Chisq)    
#1   3 -161.32                         
#2   6 -142.65  3 37.343  3.892e-08 ***


#Differences between groups (general): Shannon
full<-lme4::lmer(Shannon ~ Group + (1|Batch), sdt)
null<-lme4::lmer(Shannon ~ (1|Batch), sdt)
lrtest(null, full)
#Likelihood ratio test
#Model 1: Shannon ~ (1 | Batch)
#Model 2: Shannon ~ Group + (1 | Batch)
  #Df  LogLik Df  Chisq Pr(>Chisq)
#1   3 -4.9443                     
#2   6 -5.5971  3 1.3055     0.7278


#Differences between groups (general): Fisher
full<-lme4::lmer(Fisher ~ Group + (1|Batch), sdt)
null<-lme4::lmer(Fisher ~ (1|Batch), sdt)
lrtest(null, full)
#Likelihood ratio test
#Model 1: Fisher ~ (1 | Batch)
#Model 2: Fisher ~ Group + (1 | Batch)
  #Df  LogLik Df  Chisq Pr(>Chisq)    
#1   3 -99.028                         
#2   6 -85.546  3 26.964   5.99e-06 ***
#---

#ctr	vs. ctr+rif
filtered_sdt <- sdt %>%
  filter(Group %in% c("ctr", "ctr+rif"))
#Fisher
Ffull<-lme4::lmer(Fisher ~ Group + (1|Batch), filtered_sdt)
Fnull<-lme4::lmer(Fisher ~ (1|Batch), filtered_sdt)
F1<-lrtest(Fnull, Ffull)
 #Df  LogLik Df  Chisq Pr(>Chisq)    
#1   3 -50.473                         
#2   4 -44.399  1 12.149  0.0004912 ***
#Chao1
Cfull<-lme4::lmer(Chao1 ~ Group + (1|Batch), filtered_sdt)
Cnull<-lme4::lmer(Chao1 ~ (1|Batch), filtered_sdt)
C1<-lrtest(Cnull, Cfull)
  #Df  LogLik Df  Chisq Pr(>Chisq)    
#1   3 -80.551                        
#2   4 -72.185  1 16.731  4.307e-05 ***

#ctr	vs. ccl4
filtered_sdt <- sdt %>%
  filter(Group %in% c("ctr", "ccl4"))
#Fisher
F2full<-lme4::lmer(Fisher ~ Group + (1|Batch), filtered_sdt)
F2null<-lme4::lmer(Fisher ~ (1|Batch), filtered_sdt)
F2<-lrtest(F2null, F2full)
  #Df LogLik Df  Chisq Pr(>Chisq)  
#1   3 -41.86                       
#2   4 -40.21  1 3.2993    0.06931 .
#Chao1
C2full<-lme4::lmer(Chao1 ~ Group + (1|Batch), filtered_sdt)
C2null<-lme4::lmer(Chao1 ~ (1|Batch), filtered_sdt)
C2<-lrtest(C2null, C2full)
#Df  LogLik Df  Chisq Pr(>Chisq)   
#1   3 -72.918                        
#2   4 -69.129  1 7.5769   0.005912 **


#ctr vs.	ccl4+rif
filtered_sdt <- sdt %>%
  filter(Group %in% c("ctr", "ccl4+rif"))
#Fisher
F3full<-lme4::lmer(Fisher ~ Group + (1|Batch), filtered_sdt)
F3null<-lme4::lmer(Fisher ~ (1|Batch), filtered_sdt)
F3<-lrtest(F3null, F3full)
  #Df LogLik Df  Chisq Pr(>Chisq)  
#1   3 -39.548                       
#2   4 -37.764  1 3.5674    0.05892 .
#Chao1
C3full<-lme4::lmer(Chao1 ~ Group + (1|Batch), filtered_sdt)
C3null<-lme4::lmer(Chao1 ~ (1|Batch), filtered_sdt)
C3<-lrtest(C3null, C3full)
#Df  LogLik Df  Chisq Pr(>Chisq)   
#1   3 -69.007                         
#2   4 -65.606  1 6.8039   0.009096 **


#ctr+rif vs. ccl4
filtered_sdt <- sdt %>%
  filter(Group %in% c("ctr+rif", "ccl4"))
#Fisher
F4full<-lme4::lmer(Fisher ~ Group + (1|Batch), filtered_sdt)
F4null<-lme4::lmer(Fisher ~ (1|Batch), filtered_sdt)
F4<-lrtest(F4null, F4full)
  #Df LogLik Df  Chisq Pr(>Chisq)  
#1   3 -50.463                        
#2   4 -45.663  1 9.5997   0.001946 **
#Chao1
C4full<-lme4::lmer(Chao1 ~ Group + (1|Batch), filtered_sdt)
C4null<-lme4::lmer(Chao1 ~ (1|Batch), filtered_sdt)
C4<-lrtest(C4null, C4full)
#Df  LogLik Df  Chisq Pr(>Chisq)   
#1   3 -80.972                        
#2   4 -74.432  1 13.08  0.0002985 ***

#ctr+rif vs. ccl4+rif
filtered_sdt <- sdt %>%
  filter(Group %in% c("ctr+rif", "ccl4+rif"))
#Fisher
F5full<-lme4::lmer(Fisher ~ Group + (1|Batch), filtered_sdt)
F5null<-lme4::lmer(Fisher ~ (1|Batch), filtered_sdt)
F5<-lrtest(F5null, F5full)
  #Df LogLik Df  Chisq Pr(>Chisq)  
#1   3 -51.476                      
#2   4 -43.936  1 15.079  0.0001031 ***
#Chao1
C5full<-lme4::lmer(Chao1 ~ Group + (1|Batch), filtered_sdt)
C5null<-lme4::lmer(Chao1 ~ (1|Batch), filtered_sdt)
C5<-lrtest(C5null, C5full)
#Df  LogLik Df  Chisq Pr(>Chisq)   
#1   3 -81.164                        
#2   4 -72.420  1 17.489  2.889e-05 ***


#ccl4	vs. ccl4+rif
filtered_sdt <- sdt %>%
  filter(Group %in% c("ccl4", "ccl4+rif"))
#Fisher
F6full<-lme4::lmer(Fisher ~ Group + (1|Batch), filtered_sdt)
F6null<-lme4::lmer(Fisher ~ (1|Batch), filtered_sdt)
F6<-lrtest(F6null, F6full)
#Df  LogLik Df  Chisq Pr(>Chisq)  
#1   3 -42.340                       
#2   4 -39.967  1 4.7452    0.02938 *
#Chao1
C6full<-lme4::lmer(Chao1 ~ Group + (1|Batch), filtered_sdt)
C6null<-lme4::lmer(Chao1 ~ (1|Batch), filtered_sdt)
C6<-lrtest(C6null, C6full)
#Df  LogLik Df Chisq Pr(>Chisq)   
#1   3 -73.611                       
#2   4 -69.578  1 8.065   0.004513 **

library(dplyr)
# Create a list of test objects
fisher_tests <- list(F1, F2, F3, F4, F5, F6)
chao1_tests <- list(C1, C2, C3, C4, C5, C6)

# Create a function to extract test information and create a data frame
extract_test_info <- function(test_objects, test_type) {
  results <- lapply(test_objects, function(test) {
    data.frame(
      Test = test_type,
      LogLik_Null = test$LogLik[1],
      LogLik_Full = test$LogLik[2],
      Df = test$Df[2],
      Chisq = test$Chisq[2],
      PValue = test$`Pr(>Chisq)`[2],
      Sig = ifelse(test$`Pr(>Chisq)`[2] < 0.001, "***",
                           ifelse(test$`Pr(>Chisq)`[2] < 0.01, "**",
                                  ifelse(test$`Pr(>Chisq)`[2] < 0.05, "*", "")))
    )
  })
  return(bind_rows(results))
}

# Extract information and create results data frames
fisher_results <- extract_test_info(fisher_tests, "Fisher")
chao1_results <- extract_test_info(chao1_tests, "Chao1")
# Adjust p-value
fisher_results$adjusted_p_value <- p.adjust(fisher_results$PValue, method = "BH") 
chao1_results$adjusted_p_value <- p.adjust(chao1_results$PValue, method = "BH") 
fisher_results$Sig_adj <- ifelse(fisher_results$adjusted_p_value < 0.001, "***",
                           ifelse(fisher_results$adjusted_p_value < 0.01, "**",
                                  ifelse(fisher_results$adjusted_p_value < 0.05, "*", "")))
chao1_results$Sig_adj <- ifelse(chao1_results$adjusted_p_value < 0.001, "***",
                           ifelse(chao1_results$adjusted_p_value < 0.01, "**",
                                  ifelse(chao1_results$adjusted_p_value < 0.05, "*", "")))
#Add the comparisons
comparisons <- c(
  "ctr vs. ctr+rif",
  "ctr vs. ccl4",
  "ctr vs. ccl4+rif",
  "ctr+rif vs. ccl4",
  "ctr+rif vs. ccl4+rif",
  "ccl4 vs. ccl4+rif"
)
fisher_results$Comparisons <- comparisons
chao1_results$Comparisons <- comparisons
# Print the results data frames
print(fisher_results)
print(chao1_results)
write.csv(fisher_results, "~/Rats_HE_CCL4model/output/Tables/Fisher_models.csv")
write.csv(chao1_results, "~/Rats_HE_CCL4model/output/Tables/Chao1_models.csv")
``` 

###Alpha Diversity plots
```{r AlphaDiversity plots}
ps_2 <- readRDS("~/Library/CloudStorage/OneDrive-UniversitatdeValencia/Doctorado/Estudios/CIPF_Estudio1_2018/Rats_HE_CCL4model/output/ps_2.Rds")
newSTorder = c("ctr", "ctr+rif", "ccl4", "ccl4+rif")
metadata<- data.frame(sample_data(ps_2)) 
library(ggsignif)

Chao1<-phyloseq::estimate_richness(ps_2, measures = "Chao1")
metadata<- data.frame(sample_data(ps_2)) 
Chao1$Group<-metadata$Group
pChao1 <-ggplot(Chao1, aes(x=Group, y=Chao1))+
  geom_violin(aes(color = Group), trim = FALSE) +  
  geom_boxplot(width = 0.2) +
  theme_classic() + theme(strip.background = element_blank(), axis.text.x.bottom = element_text(angle = -90), axis.text.x = element_text(size = 16), axis.title.y = element_text(size = 16)) +
      geom_signif(stat="identity",
               data=data.frame(x=c(1, 2, 3), xend=c(2, 3, 4), y=c(425, 430, 435), annotation=c("***", " *** ", "**")),
               aes(x=x,xend=xend, y=y, yend=y, annotation=annotation)) +
   geom_signif(comparisons=list(c("ctr", "ccl4")), annotations="**",  y_position = 440, tip_length = 0, vjust=0.4)  +
   geom_signif(comparisons=list(c("ctr", "ccl4+rif")), annotations="**",  y_position = 480, tip_length = 0, vjust=0.4) +
   geom_signif(comparisons=list(c("ctr+rif", "ccl4+rif")), annotations="***",  y_position = 460, tip_length = 0, vjust=0.4) 
#Re-arranging the order of the figure, controls first
pChao1$data$Group <- as.character(pChao1$data$Group)
pChao1$data$Group <- factor(pChao1$data$Group, levels=newSTorder)
pChao1


Shannon<-phyloseq::estimate_richness(ps_2, measures = "shannon")
metadata<- data.frame(sample_data(ps_2)) 
Shannon$Group<-metadata$Group
pShannon <-ggplot(Shannon, aes(x=Group, y=Shannon))+
  geom_violin(aes(color = Group), trim = FALSE) +  
  geom_boxplot(width = 0.2) +
  theme_classic() + theme(strip.background = element_blank(), axis.text.x.bottom = element_text(angle = -90), axis.text.x = element_text(size = 16), axis.title.y = element_text(size = 16))  +
  ggtitle("p=0.73;Df=3;Chi-sq=1.31")  # Add title

#Re-arranging the order of the figure, controls first
pShannon$data$Group <- as.character(pShannon$data$Group)
pShannon$data$Group <- factor(pShannon$data$Group, levels=newSTorder)
pShannon

Fisher<-phyloseq::estimate_richness(ps_2, measures = "fisher")
metadata<- data.frame(sample_data(ps_2)) 
Fisher$Group<-metadata$Group
pFisher <-ggplot(Fisher, aes(x=Group, y=Fisher))+
  geom_violin(aes(color = Group), trim = FALSE) +  
  geom_boxplot(width = 0.2) +
  theme_classic() + theme(strip.background = element_blank(), axis.text.x.bottom = element_text(angle = -90), axis.text.x = element_text(size = 16), axis.title.y = element_text(size = 16))  +
      geom_signif(stat="identity",
               data=data.frame(x=c(1, 2, 3), xend=c(2, 3, 4), y=c(55, 57, 59), annotation=c("**", " ** ", "*")),
               aes(x=x,xend=xend, y=y, yend=y, annotation=annotation)) +
   geom_signif(comparisons=list(c("ctr+rif", "ccl4+rif")), annotations="***",  y_position = 60, tip_length = 0, vjust=0.4) 
#Re-arranging the order of the figure, controls first
pFisher$data$Group <- as.character(pFisher$data$Group)
pFisher$data$Group <- factor(pFisher$data$Group, levels=newSTorder)
pFisher

ggarrange(pShannon, pFisher, pChao1, common.legend = TRUE, legend="bottom",nrow = 1)


all_ASVs_boxplots <- ggarrange(pShannon, pChao1, pFisher,common.legend = TRUE, legend="bottom",nrow = 1)

# Modify the legend order in the updated plot
all_ASVs_boxplots <- ggsave(filename = "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Figures/AphaDiv_ASVs_padj.pdf", width = 12, height = 8)
```
