==> R/4_BatchEffectRemoval_and_Betadiv.Rmd <==
---
title: "4_BatchEffectRemoval_and_Betadiv"
output: html_document
date: "2023-04-27"
---
###Open the data
```{r Data}
#ps_2 contains raw data removing ASVs that are all zeros (prune_taxa(taxa_sums(ps_1) > 0, ps_1))
ps_2 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_2.Rds")
library(readxl)
metadata <- read_excel("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/metadata.xlsx")
View(metadata)
```

###Only zeros data removal
```{r Log and zeros removal}
#Raw data
rawDada2_Species <- as.data.frame(otu_table(ps_2))
rawDada2_Species <- as.data.frame(rawDada2_Species)
rawDada2_Species <- rawDada2_Species[,metadata$SampleID]

# Delete rows with all values = 0
rawDada2_Species <- rawDada2_Species[rowSums(rawDada2_Species) != 0,]

# Add 1 count to avoid log problems
rawDada2_Species <- rawDada2_Species + 1

total = colSums(rawDada2_Species)

```

###Boxplots
```{r Boxplots}
boxplot(rawDada2_Species, las = 2)
boxplot(log(rawDada2_Species), las = 2)
``` 

###Low counts: removing zeros
```{r Zeros}
# Note: we replaced all zero values by 1 to avoid log errors
zeros <- t(apply(rawDada2_Species, 1, function(i){
  do.call(cbind, lapply(unique(metadata$Group), function(x){
    sum(i[c(metadata$Group == x)] == 1)
  }))
}))
colnames(zeros) = unique(metadata$Group)
zeros <- data.frame(zeros[rownames(rawDada2_Species),])

bp <- do.call(cbind, lapply(1:10, function(y){
  logic_z <- data.frame(zeros <= y)
  logic_z$sum <- rowSums(logic_z)
  sum(logic_z$sum >= 1)
}))
colnames(bp)= 1:10

{my_bar <- barplot(bp, main = "filter in at least 1 group <= number of zeros", 
        xlab = "Number of zeros", ylab = "number of final Species",
        ylim = c(0,max(bp)+20))
text(my_bar, bp+10 , paste("n: ", bp, sep="") ,cex=1) }
``` 


###Filtering: removing zeros
```{r Filter}
logic_z <- data.frame(zeros <= 3)
logic_z$sum <- rowSums(logic_z)
rawDada2_Species_f = rawDada2_Species[logic_z$sum >= 1,]

# Data in percentage (%)
total = colSums(rawDada2_Species_f)


boxplot(rawDada2_Species_f, las = 2)
boxplot(log(rawDada2_Species_f+1), las = 2)
```

##New Phyloseq object without zeros - no transformations data

```{r new phyloseq no zeros} 
# Remove samples with less than MINREADS from phyloseq object
ps_2 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_2.Rds")


OTU_data_0 <- rawDada2_Species[logic_z$sum >= 1,]
OTU_data_0$names <-row.names(OTU_data_0)
tax_table_0 <-as.data.frame(as.matrix(tax_table(ps_2)))
tax_table_0$names <-row.names(tax_table_0)

NewTAX <- semi_join(tax_table_0, OTU_data_0, by = join_by(names))
NewTAX <- subset(NewTAX, select = -c(names))
OTU_data_0 <- rawDada2_Species[logic_z$sum >= 1,]

OTU = otu_table((OTU_data_0-1), taxa_are_rows = TRUE)
TAX = tax_table(as.matrix(NewTAX))
SAMPLE <- sample_data(metadata)
phylo_tree <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/phylo_tree.rds")
TREE = phylo_tree$tree
  
#Como me daba error hetenido que especificar qué datos son qué columnas
rownames(SAMPLE) <-metadata$SampleID


# merge the data
ps_0 <- phyloseq(OTU, TAX, SAMPLE, TREE)
ps_0
saveRDS(ps_0, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_0.Rds")
```

###PCoAs: before removing batch effect per group
Without log transformation
```{r PCoA representation}
library(tibble)
library(ggplot2)
library(forcats)
# start with your phyloseq object
ps.taxa.sub <- subset_samples(ps_0)
phyloseq.obj <- ps.taxa.sub

pco.data <- phyloseq.obj@otu_table %>%
    t() %>%
    as.data.frame() %>%
    # bray-curtis dissimilarity
    vegan::vegdist(method = 'bray',
                   diag = TRUE, upper = TRUE) %>%
    # multidimensional scaling, aka PCoA
    stats::cmdscale(k = 2, eig = TRUE) 

# get group labels 
cohort.labels <- phyloseq.obj@sam_data %>%
    data.frame() %>%
    as_tibble() %>%
    # replace below with your own code, e.g. Group and other variables
    select(SampleID, Group) %>%
    # change labels, format: desired name = 'existing name'
    mutate_at(vars(Group), ~ fct_recode(.,
                                         `ccl4` = 'CCL4',
                                         `ccl4+rif` = 'CCL4+rif',
                                        `ctr+rif` = 'Ctr+rif', 
                                         `ctr` = 'Control')) %>%
    # change order
    mutate_at(vars(Group), ~ fct_relevel(., 'Control', 'Ctr+rif','CCL4','CCL4+rif'))
# add cohort labels/metadata for plotting
plot.points <- pco.data$points %>%
    as.data.frame() %>%
    rownames_to_column('SampleID') %>%
    as_tibble() %>%
    left_join(cohort.labels)

# save first and second components to label % variation explained on plots
axes <- sprintf(fmt='%.2f',
                (pco.data$eig[1:2]/sum(pco.data$eig[pco.data$eig > 0])) * 100) 

# plot 
plot.points %>%
    ggplot(aes(x = V1, y = V2), color = Group) + stat_ellipse(aes(color = Group)) +
    geom_point(aes(color = Group), size=3, shape=21, alpha=0.8) +
    ggsci::scale_fill_npg() +
    labs(
        size = 'Point Type', 
        x = paste0('PCo 1 [', axes[1], '%]'), 
        y = paste0('PCo 2 [', axes[2], '%]')) + 
    guides(size = FALSE, shape = FALSE) +
    ggpubr::theme_pubr() +
    ggpubr::labs_pubr() +
    theme(legend.position = c(0.59, 0.9))
```

###PCoAs: before removing batch effect per batch 
Without log transformation
```{r PCoA representation}
# start with your phyloseq object
ps.taxa.sub <- subset_samples(ps_0)
phyloseq.obj <- ps.taxa.sub

pco.data <- phyloseq.obj@otu_table %>%
    t() %>%
    as.data.frame() %>%
    # bray-curtis dissimilarity
    vegan::vegdist(method = 'bray',
                   diag = TRUE, upper = TRUE) %>%
    # multidimensional scaling, aka PCoA
    stats::cmdscale(k = 2, eig = TRUE) 
# get group labels 
cohort.labels <- phyloseq.obj@sam_data %>%
    data.frame() %>%
    as_tibble()

# add cohort labels/metadata for plotting
plot.points <- pco.data$points %>%
    as.data.frame() %>%
    rownames_to_column('SampleID') %>%
    as_tibble() %>%
    left_join(cohort.labels)

# save first and second components to label % variation explained on plots
axes <- sprintf(fmt='%.2f',
                (pco.data$eig[1:2]/sum(pco.data$eig[pco.data$eig > 0])) * 100) 

# plot 
plot.points %>%
    ggplot(aes(x = V1, y = V2), color = Batch) + stat_ellipse(aes(color = Batch)) +
    geom_point(aes(color = Batch), size=3, shape=21, alpha=0.8) +
    ggsci::scale_fill_npg() +
    labs(
        size = 'Point Type', 
        x = paste0('PCo 1 [', axes[1], '%]'), 
        y = paste0('PCo 2 [', axes[2], '%]')) + 
    guides(size = FALSE, shape = FALSE) +
    ggpubr::theme_pubr() +
    ggpubr::labs_pubr() +
    theme(legend.position = c(0.59, 0.9))
```


###PCAs: before removing batch effect
With log transformation
```{r before removing batch effect}
library(mixOmics)
pca = pca(t(log10(rawDada2_Species_f)), ncomp = 2, center = T, scale = F)
plotIndiv(pca, group = metadata$Group, legend = T, title = "log10")

plotIndiv(pca, group = metadata$Batch, legend = T, title = "Log10")

plot(pca$loadings$X[,1], pca$loadings$X[,2], xlab = "Comp1", ylab = "Comp2",
     main = "loading plot")
```

###Limma package- function: removebatcheffect
```{r Batch Correction}
#library(limma)
#y <- removeBatchEffect((log10(rawDada2_Species_f)), batch = metadata$Batch)
#pca = pca(t(y), ncomp = 2, center = T, scale = F)
#plotIndiv(pca, group = metadata$Group, legend = T, title = "log10 + removelima") 
#plotIndiv(pca, group = metadata$Batch, legend = T, title = "log10 + removelima") 
```

###ARSyN: MultiBaC package
```{r Batch Correction}
library(MultiBaC)

# transformation
otu <- log10(rawDada2_Species_f)

# Batch correction
my_mbac <- createMbac (inputOmics = list(otu_T5 = otu[,metadata$Batch == "T5"], 
                                         otu_T7 = otu[,metadata$Batch == "T7"]),
                       batchFactor = c("T5","T7"),
                       experimentalDesign = list(
                       "T5" = metadata$Group[metadata$Batch == "T5"],
                       "T7" = metadata$Group[metadata$Batch == "T7"]),
                       omicNames = "otu")
arsyn_1 <- ARSyNbac(my_mbac, modelName = "16S", Variability = 0.95, 
                 batchEstimation = TRUE, Interaction = FALSE, beta=2)     

otu_corrected <- t(cbind(MultiAssayExperiment::assay(arsyn_1$CorrectedData$T5),                    MultiAssayExperiment::assay(arsyn_1$CorrectedData$T7)))


```
pca = pca(otu_clr_corrected, ncomp = 2, center = T, scale = F)
plotIndiv(pca, group = metadata$Group, legend = T, title = "Log10 + ARSyN")
We keep the transformed data with ARSyN because it transforms the data in a way that the batch effect is removed.

##New Phyloseq object with transformed data(ARSyN)
```{r new phyloseq} 
OTU_data_t <- as.data.frame(t(otu_corrected)+6)
#NewTax from ps_0. Needs to be run first
NewTAX 

OTU = otu_table(OTU_data_t, taxa_are_rows = TRUE)
TAX = tax_table(as.matrix(NewTAX))
SAMPLE <- sample_data(metadata)
phylo_tree <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/phylo_tree.rds")
TREE = phylo_tree$tree

rownames(SAMPLE) <-metadata$SampleID
 
 # merge the data
 ps <- phyloseq(OTU, TAX, SAMPLE, TREE)
 ps
saveRDS(ps,"~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps.Rds")
```

###PCoAs: after removing batch effect per group
```{r PCoA representation}
library(tibble)
library(ggplot2)
library(forcats)
# start with your phyloseq object
ps.taxa.sub <- subset_samples(ps)
phyloseq.obj <- ps.taxa.sub

pco.data <- phyloseq.obj@otu_table %>%
    t() %>%
    as.data.frame() %>%
    # bray-curtis dissimilarity
    vegan::vegdist(method = 'bray',
                   diag = TRUE, upper = TRUE) %>%
    # multidimensional scaling, aka PCoA
    stats::cmdscale(k = 2, eig = TRUE) 

# get group labels 
cohort.labels <- phyloseq.obj@sam_data %>%
    data.frame() %>%
    as_tibble() %>%
    # change order
    mutate_at(vars(Group), ~ fct_relevel(., 'ctr', 'ctr+rif','ccl4','ccl4+rif'))

# add cohort labels/metadata for plotting
plot.points <- pco.data$points %>%
    as.data.frame() %>%
    rownames_to_column('SampleID') %>%
    as_tibble() %>%
    left_join(cohort.labels)

# save first and second components to label % variation explained on plots
axes <- sprintf(fmt='%.2f',
                (pco.data$eig[1:2]/sum(pco.data$eig[pco.data$eig > 0])) * 100) 

# plot 
plot<-plot.points %>%
    ggplot(aes(x = V1, y = V2, color = Group)) +
    stat_ellipse(aes(color = Group)) +
    geom_point(size = 3, shape = 21, alpha = 0.8) +
    labs(
        size = 'Point Type',
        x = paste0('PCo 1 [', axes[1], '%]'),
        y = paste0('PCo 2 [', axes[2], '%]')
    ) +
    guides(size = FALSE, shape = FALSE) +
    theme_pubr() +
    labs_pubr() +
    theme(
        legend.position = "bottom",  # Legend at the bottom
        legend.text = element_text(size = 20),  # Set legend text size
        legend.title = element_blank()  # Remove legend title
    )+scale_colour_discrete(name = "Group", labels = c("Control" ,"Control+Rifaximin","CCl4","CCl4+Rifaximin"))
plot
plot <- ggsave(filename = "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Figures/BetaDiv_PCoA_log10_ARSyN_Group_removing3zeros.pdf", width = 9, height = 8)

```

###PCoAs: after removing batch effect per batch 
```{r PCoA representation}
# start with your phyloseq object
ps.taxa.sub <- subset_samples(ps)
phyloseq.obj <- ps.taxa.sub

pco.data <- phyloseq.obj@otu_table %>%
    t() %>%
    as.data.frame() %>%
    # bray-curtis dissimilarity
    vegan::vegdist(method = 'bray',
                   diag = TRUE, upper = TRUE) %>%
    # multidimensional scaling, aka PCoA
    stats::cmdscale(k = 2, eig = TRUE) 
# get group labels 
cohort.labels <- phyloseq.obj@sam_data %>%
    data.frame() %>%
    as_tibble()

# add cohort labels/metadata for plotting
plot.points <- pco.data$points %>%
    as.data.frame() %>%
    rownames_to_column('SampleID') %>%
    as_tibble() %>%
    left_join(cohort.labels)

# save first and second components to label % variation explained on plots
axes <- sprintf(fmt='%.2f',
                (pco.data$eig[1:2]/sum(pco.data$eig[pco.data$eig > 0])) * 100) 

# plot 
plotbatch<-plot.points %>%
    ggplot(aes(x = V1, y = V2), color = Batch) + stat_ellipse(aes(color = Batch)) +
    geom_point(aes(color = Batch), size=3, shape=21, alpha=0.8) +
    ggsci::scale_fill_npg() +
    labs(
        size = 'Point Type', 
        x = paste0('PCo 1 [', axes[1], '%]'), 
        y = paste0('PCo 2 [', axes[2], '%]')) + 
    guides(size = FALSE, shape = FALSE) +
    ggpubr::theme_pubr() +
    ggpubr::labs_pubr() +
    theme(legend.position = c(0.59, 0.9))
plotbatch

plotbatch<- ggsave(filename = "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Figures/BetaDiv_PCoA_log10_ARSyN_Batch_removing3zeros.pdf", width = 9, height = 8)
```

#PCoAs changing aesthetic

```{r PCoA representation after batch effect removal 2}
BC_dist<- phyloseq::distance(ps, method="bray", weighted=T)
ordination<- ordinate(ps,
                      method="PCoA", distance= BC_dist)
plot_ordination(ps, ordination, shape= "Group")+ 
    theme(aspect.ratio=1)+
    geom_point(size=3, aes(color= Group))+
    labs(title = "Bray-Curtis dissimilarity",tag= "A)")+
    theme_bw()+
    theme(text = element_text(size=16))+
    labs(colour = "Group")+
    xlab(paste0("PCo 1 [", round(ordination$values[1,2]*100, digits = 2), "%]"))+
    ylab(paste0("PCo 2 [", round(ordination$values[2,2]*100, digits = 2), "%]"))
```

#PCoAs Centroid estimation

```{r Centroid estimation}
##Bray curtis estimation
bray_matrix<- ps@otu_table %>%
  t() %>%
  as.data.frame() %>%
  # bray-curtis dissimilarity
  vegan::vegdist(method = 'bray',
                 diag = TRUE, upper = TRUE)
metadata<-as.data.frame(metadata)
rownames(metadata)<-metadata$SampleID
Group<-merge(as.data.frame(as.matrix(bray_matrix)), metadata, by="row.names")
Group<-as.vector(rownames(as.data.frame(as.matrix(bray_matrix))))

df <- metadata
df<-df[match(Group, df$SampleID),]

## Calculate multivariate dispersion or distance to the centroid
permdisp <- vegan::betadisper(bray_matrix, df$Group, "centroid")

##Plot
plot(permdisp, main="PCoA")
boxplot(permdisp, main="Distance to centroids")
```

#3D PCoAs
```{r PCoA representation after batch effect removal 2}
library(rgl)
library(knitr)
library(tibble)
library(ggplot2)
library(dplyr)
# start with your phyloseq object
ps.taxa.sub <- subset_samples(ps)
phyloseq.obj <- ps.taxa.sub

pco.data <- phyloseq.obj@otu_table %>%
    t() %>%
    as.data.frame() %>%
    # bray-curtis dissimilarity
    vegan::vegdist(method = 'bray',
                   diag = TRUE, upper = TRUE) %>%
    # multidimensional scaling, aka PCoA
    stats::cmdscale(k = 3, eig = TRUE) 
# get group labels 
cohort.labels <- phyloseq.obj@sam_data %>%
    data.frame() %>%
    as_tibble()

# add cohort labels/metadata for plotting
plot.points <- pco.data$points %>%
    as.data.frame() %>%
    rownames_to_column('SampleID') %>%
    as_tibble() %>%
    left_join(cohort.labels)
axes <- sprintf(fmt='%.2f',
                (pco.data$eig[1:3]/sum(pco.data$eig[pco.data$eig > 0])) * 100)

# This is to output a rgl plot in a rmarkdown document. Note that you must add webgl=TRUE, results='hide' in the chunck header
library(knitr)
knit_hooks$set(webgl = hook_webgl)

# Data: the iris data is provided by R
data <- plot.points

# Add a new column with color
mycolors <- c('royalblue1', 'lightgreen', 'pink', 'orange')
data$color <- mycolors[ as.factor(data$Group) ]

# Plot
par(mar=c(0,0,0,0))
p<-plot3d( 
    x=data$V1, y=data$V2, z=data$V3, 
    col = data$color, size = 3,
    type = 's', radius=1/500,
    xlab=paste0('PCo 1 [', axes[1], '%]'), ylab=paste0('PCo 2 [', axes[2], '%]'), zlab=paste0('PCo 3 [', axes[3], '%]'))

htmlwidgets::saveWidget(rglwidget(width = 720, height = 720), 
                        file = "~/Documents/PCoA_Groups_3D_Spheres.html",
                        libdir = "libs",
                        selfcontained = FALSE
)

# Plot
par(mar=c(0,0,0,0))
p<-plot3d( 
    x=data$V1, y=data$V2, z=data$V3, 
    col = data$color, size = 20,
    type = 'p',
    xlab=paste0('PCo 1 [', axes[1], '%]'), ylab=paste0('PCo 2 [', axes[2], '%]'), zlab=paste0('PCo 3 [', axes[3], '%]'))

htmlwidgets::saveWidget(rglwidget(width = 720, height = 720), 
                        file = "~/Documents/PCoA_Groups_3D_Points.html",
                        libdir = "libs",
                        selfcontained = FALSE
)

```
