==> R/5_MetadeconfoundR_taxa.Rmd <==
---
title: "5_MetadeconfoundR"
author: "Lola"
date: "5/2/2023"
output: html_document
---
#Data
```{r MetadeconfoundR heatmaps}
library(devtools)
install_github("TillBirkner/metadeconfoundR")

ps_0 <- readRDS("/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/ps_0.Rds")
library(readxl)
metadata <- read_excel("/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/metadata.xlsx")
View(metadata)
```

```{r Changing names ASVs}
library(tidyr)
library(tibble)
library(purrr)
require(phyloseq)
require(tidyverse)
require(magrittr)


get_latest_annotation <- function(phyloseq_obj) {
  tax_table <- phyloseq_obj@tax_table %>%
    as.data.frame() %>%
    rownames_to_column('ASV') %>%
    as_tibble() %>%
    tidyr::gather('Rank', 'Value', -ASV) %>%
    nest(data = -ASV) %>%
    mutate(TaxaID = map_chr(data, function(x) {
      x %>%
        dplyr::filter(!is.na(Value)) %>%
        magrittr::use_series(Value) %>%
        tail(1)
    })) %>%
    mutate(TaxaUp = map_chr(data, function(x) {
      x %>%
        dplyr::filter(!is.na(Value)) %>%
        magrittr::use_series(Value) %>%
        tail(2) %>% head(1)
    })) %>%
    mutate(TaxaID = paste(ASV, TaxaID)) %>%
    dplyr::select(-c(data, TaxaUp))
  
  return(tax_table)
}

##Usage latest_annotations <- get_latest_annotation(phyloseq_obj)
latest_annotations <- get_latest_annotation(ps_0)
latest_annotations<-as.data.frame(latest_annotations)
short_names<-as.data.frame(otu_table(ps_0))
rownames(short_names)<-latest_annotations$TaxaID
```

```{r Changing names ASVs long}
# generate a vector containing the full taxonomy path for all OTUs
wholetax <- do.call(paste, c(as.data.frame(tax_table(ps_0))
                             [c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species")], 
                             sep = "__"))  # to distinguish from "_" within tax ranks

# turn the otu_table into a data.frame
otu_export <- as.data.frame(t(otu_table(ps_0)))
tmp <- names(otu_export)

# paste wholetax and OTU_ids together
for(i in 1:length(tmp)){
    names(tmp)[i] = paste(wholetax[i], tmp[i], sep = "__")
}

# overwrite old names
names(otu_export) <- names(tmp)
otu_export<-t(otu_export)

head(otu_export)
```

##MetadeconfoundR
```{r MetadeconfoundR heatmaps}
library(metadeconfoundR)
library(dplyr)
ps.rel<-microbiome::transform(ps_0, "compositional")
latest_annotations <- get_latest_annotation(ps.rel)
latest_annotations<-as.data.frame(latest_annotations)
short_names<-as.data.frame(otu_table(ps.rel))
rownames(short_names)<-latest_annotations$TaxaID
Meta_Species <- t(short_names) 

metadata_R <- subset(metadata, select = -c(SampleID, Week, FisabioID, GABAa1, GABAa5, Ymaze_trialstolearn, OLM, IL1_b))
metadata_R <- metadata_R %>% mutate(Batch = ifelse(Batch=="T5",1,0))
metadata_R <- metadata_R %>% mutate(Antibiotic = ifelse(Antibiotic=="Yes",1,0))
metadata_R <- metadata_R %>% mutate(Treated_CCL = ifelse(Treated_CCL=="Yes",1,0))
metadata_R <- as.data.frame(metadata_R)
rownames(metadata_R) <- (metadata$SampleID)

# check correct ordering
all(rownames(metadata_R) == rownames(Meta_Species))
## [1] TRUE
all(order(rownames(metadata_R)) == order(rownames(Meta_Species)))
## [1] TRUE
Meta_Species<- as.data.frame(Meta_Species)
Output1 <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                 metaMat = as.data.frame(metadata_R), nnodes = 14)
                                 
Output2_batch <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                     metaMat = as.data.frame(metadata_R), nnodes = 14, randomVar = list("+ (1|Batch)",
                                                                                        c("Batch")))
View(Output1)
View(Output2_batch)

left <- BuildHeatmap(Output1)
right <- BuildHeatmap(Output2_batch)
plot(left)
plot(right)
saveRDS(Output1, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output1_metadec.Rds")
saveRDS(Output2_batch, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output2_batch_all.Rds")

```

##MetadeconfoundR genus
```{r MetadeconfoundR heatmaps}
ps_0_genus <- tax_glom(ps_0, taxrank = 'Genus', NArm = FALSE)
ps.rel<-microbiome::transform(ps_0_genus, "compositional")
latest_annotations <- get_latest_annotation(ps.rel)
latest_annotations<-as.data.frame(latest_annotations)
short_names<-as.data.frame(otu_table(ps.rel))
rownames(short_names)<-latest_annotations$TaxaID
Meta_Species <- t(short_names) 

metadata_R <- subset(metadata, select = -c(SampleID, Week, FisabioID, GABAa1, GABAa5, Ymaze_trialstolearn, OLM, IL1_b))
metadata_R <- metadata_R %>% mutate(Batch = ifelse(Batch=="T5",1,0))
metadata_R <- metadata_R %>% mutate(Antibiotic = ifelse(Antibiotic=="Yes",1,0))
metadata_R <- metadata_R %>% mutate(Treated_CCL = ifelse(Treated_CCL=="Yes",1,0))
metadata_R <- as.data.frame(metadata_R)
rownames(metadata_R) <- (metadata$SampleID)

# check correct ordering
all(rownames(metadata_R) == rownames(Meta_Species))
## [1] TRUE
all(order(rownames(metadata_R)) == order(rownames(Meta_Species)))
## [1] TRUE
Meta_Species<- as.data.frame(Meta_Species)
                                 
Output2_batch <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                     metaMat = as.data.frame(metadata_R), nnodes = 14, randomVar = list("+ (1|Batch)",
                                                                                        c("Batch")))

View(Output2_batch)


right <- BuildHeatmap(Output2_batch)

plot(right)
saveRDS(Output2_batch, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output2_batch_allgenus.Rds")
```

##MetadeconfoundR ctr vs ctr+rif
```{r MetadeconfoundR heatmaps}
library(metadeconfoundR)
library(dplyr)
library(phyloseq)
ps.rel<-ps_0 %>% microbiome::transform("compositional") %>% subset_samples(Group %in% c("ctr", "ctr+rif"))
latest_annotations <- get_latest_annotation(ps.rel)
latest_annotations<-as.data.frame(latest_annotations)
short_names<-as.data.frame(otu_table(ps.rel))
rownames(short_names)<-latest_annotations$TaxaID
Meta_Species <- t(short_names) 

metadata_R <- subset(data.frame(sample_data(ps.rel)), select = -c(Treated_CCL, SampleID, Week, FisabioID, GABAa1, GABAa5, Ymaze_trialstolearn, OLM, IL1_b, IL_6, TNF_a, IFN_gamma, IL_15, IL_17, IL_10,  IL_4, TGF_b,  CCL2, CCL5, CCL20, CX3CL1,  CCR2,  CCR5, CX3CR1, Occludine, Learning_index_Rmaze))
metadata_R <- metadata_R %>% mutate(Batch = ifelse(Batch=="T5",1,0))
metadata_R <- metadata_R %>% mutate(Antibiotic = ifelse(Antibiotic=="Yes",1,0))
metadata_R <- as.data.frame(metadata_R)

# check correct ordering
all(rownames(metadata_R) == rownames(Meta_Species))
## [1] TRUE
all(order(rownames(metadata_R)) == order(rownames(Meta_Species)))
## [1] TRUE
Meta_Species<- as.data.frame(Meta_Species)
Output1_ctrvsctrrif <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                 metaMat = as.data.frame(metadata_R), nnodes = 14)
                                 
Output2_batch_ctrvsctrrif <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                     metaMat = as.data.frame(metadata_R), nnodes = 14, randomVar = list("+ (1|Batch)",
                                                                                        c("Batch")))
View(Output1_ctrvsctrrif)
View(Output2_batch_ctrvsctrrif)

left <- BuildHeatmap(Output1_ctrvsctrrif)
right <- BuildHeatmap(Output2_batch_ctrvsctrrif)
plot(left)
plot(right)
saveRDS(Output1_ctrvsctrrif, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output1_metadec_ctrvsctrrif.Rds")
saveRDS(Output2_batch_ctrvsctrrif, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output2_batch_asRandVar_ctrvsctrrif.Rds")

```

##MetadeconfoundR ctr vs ccl4
```{r MetadeconfoundR heatmaps}
library(metadeconfoundR)
library(dplyr)
library(phyloseq)
ps.rel3<-ps_0 %>% microbiome::transform("compositional") %>% subset_samples(Group %in% c("ctr", "ccl4"))
latest_annotations <- get_latest_annotation(ps.rel3)
latest_annotations<-as.data.frame(latest_annotations)
short_names<-as.data.frame(otu_table(ps.rel3))
rownames(short_names)<-latest_annotations$TaxaID
Meta_Species <- t(short_names) 

metadata_R <- subset(data.frame(sample_data(ps.rel3)), select = -c(Antibiotic, SampleID, Week, FisabioID, GABAa1, GABAa5, Ymaze_trialstolearn, Ammonia))
metadata_R <- metadata_R %>% mutate(Batch = ifelse(Batch=="T5",1,0))
metadata_R <- metadata_R %>% mutate(Treated_CCL = ifelse(Treated_CCL=="Yes",1,0))
metadata_R <- as.data.frame(metadata_R)

# check correct ordering
all(rownames(metadata_R) == rownames(Meta_Species))
## [1] TRUE
all(order(rownames(metadata_R)) == order(rownames(Meta_Species)))
## [1] TRUE
Meta_Species<- as.data.frame(Meta_Species)
Output1_ctrvsccl4 <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                 metaMat = as.data.frame(metadata_R), nnodes = 14)
                                 
Output2_batch_ctrvsccl4 <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                     metaMat = as.data.frame(metadata_R), nnodes = 14, randomVar = list("+ (1|Batch)",
                                                                                        c("Batch")))
View(Output1_ctrvsccl4)
View(Output2_batch_ctrvsccl4)

left <- BuildHeatmap(Output1_ctrvsccl4)
right <- BuildHeatmap(Output2_batch_ctrvsccl4)
plot(left)
plot(right)
saveRDS(Output1_ctrvsccl4, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output1_ctrvsccl4.Rds")
saveRDS(Output2_batch_ctrvsccl4, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output2_batch_ctrvsccl4.Rds")

```

##MetadeconfoundR ctr vs ccl4+rif
```{r MetadeconfoundR heatmaps}
library(metadeconfoundR)
library(dplyr)
ps.rel4<-ps_0 %>% microbiome::transform("compositional") %>% subset_samples(Group %in% c("ctr", "ccl4+rif"))
latest_annotations <- get_latest_annotation(ps.rel4)
latest_annotations<-as.data.frame(latest_annotations)
short_names<-as.data.frame(otu_table(ps.rel4))
rownames(short_names)<-latest_annotations$TaxaID
Meta_Species <- t(short_names) 

metadata_R <- subset(data.frame(sample_data(ps.rel4)), select = -c(SampleID, Week, FisabioID, GABAa1, GABAa5, Ymaze_trialstolearn, OLM, IFN_gamma, IL_15, IL_17, IL_10,  IL_4, TGF_b,  CCL2, CCL5, CCL20, CX3CL1,  CCR2,  CCR5, CX3CR1, Occludine, Learning_index_Rmaze, IL_6, IL1_b, TNF_a, Ammonia, Ambulatory_Counts, Vertcal_Counts, Stereotipic_Counts, Average_velocity))
metadata_R <- metadata_R %>% mutate(Batch = ifelse(Batch=="T5",1,0))
metadata_R <- metadata_R %>% mutate(Treated_CCL = ifelse(Treated_CCL=="Yes",1,0))
metadata_R <- metadata_R %>% mutate(Antibiotic = ifelse(Antibiotic=="Yes",1,0))
metadata_R <- as.data.frame(metadata_R)

# check correct ordering
all(rownames(metadata_R) == rownames(Meta_Species))
## [1] TRUE
all(order(rownames(metadata_R)) == order(rownames(Meta_Species)))
## [1] TRUE
Meta_Species<- as.data.frame(Meta_Species)
Output1_ctrvsccl4rif <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                 metaMat = as.data.frame(metadata_R), nnodes = 14)
                                 
Output2_batch_ctrvsccl4rif <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                     metaMat = as.data.frame(metadata_R), nnodes = 14, randomVar = list("+ (1|Batch)",
                                                                                        c("Batch")))
View(Output1_ctrvsccl4rif)
View(Output2_batch_ctrvsccl4rif)

left <- BuildHeatmap(Output1_ctrvsccl4rif)
right <- BuildHeatmap(Output2_batch_ctrvsccl4rif)
plot(left)
plot(right)
saveRDS(Output1_ctrvsccl4rif, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output1_ctrvsccl4rif.Rds")
saveRDS(Output2_batch_ctrvsccl4rif, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output2_batch_ctrvsccl4rif.Rds")

```

##MetadeconfoundR ccl4 vs ccl4+rif
```{r MetadeconfoundR heatmaps}
library(metadeconfoundR)
library(dplyr)
ps.rel5<-ps_0 %>% microbiome::transform("compositional") %>% subset_samples(Group %in% c("ccl4", "ccl4+rif"))
latest_annotations <- get_latest_annotation(ps.rel5)
latest_annotations<-as.data.frame(latest_annotations)
short_names<-as.data.frame(otu_table(ps.rel5))
rownames(short_names)<-latest_annotations$TaxaID
Meta_Species <- t(short_names) 

metadata_R <- subset(data.frame(sample_data(ps.rel5)), select = -c(Treated_CCL, SampleID, Week, FisabioID, GABAa1, GABAa5, Ymaze_trialstolearn, OLM, IL1_b, Ammonia))
metadata_R <- metadata_R %>% mutate(Batch = ifelse(Batch=="T5",1,0))
metadata_R <- metadata_R %>% mutate(Antibiotic = ifelse(Antibiotic=="Yes",1,0))
metadata_R <- as.data.frame(metadata_R)

# check correct ordering
all(rownames(metadata_R) == rownames(Meta_Species))
## [1] TRUE
all(order(rownames(metadata_R)) == order(rownames(Meta_Species)))
## [1] TRUE
Meta_Species<- as.data.frame(Meta_Species)
Output1_ccl4vsccl4rif <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                 metaMat = as.data.frame(metadata_R), nnodes = 14)
                                 
Output2_batch_ccl4vsccl4rif <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                     metaMat = as.data.frame(metadata_R), nnodes = 14, randomVar = list("+ (1|Batch)",
                                                                                        c("Batch")))
View(Output1_ccl4vsccl4rif)
View(Output2_batch_ccl4vsccl4rif)

left <- BuildHeatmap(Output1_ccl4vsccl4rif)
right <- BuildHeatmap(Output2_batch_ccl4vsccl4rif)
plot(left)
plot(right)
saveRDS(Output1_ccl4vsccl4rif, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output1_ccl4vsccl4rif.Rds")
saveRDS(Output2_batch_ccl4vsccl4rif, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output2_batch_ccl4vsccl4rif.Rds")

```

##MetadeconfoundR ccl4+rif vs ctr+rif
```{r MetadeconfoundR heatmaps}
library(metadeconfoundR)
library(dplyr)
ps.rel6<-ps_0 %>% microbiome::transform("compositional") %>% subset_samples(Group %in% c("ctr+rif", "ccl4+rif"))
latest_annotations <- get_latest_annotation(ps.rel6)
latest_annotations<-as.data.frame(latest_annotations)
short_names<-as.data.frame(otu_table(ps.rel6))
rownames(short_names)<-latest_annotations$TaxaID
Meta_Species <- t(short_names) 

metadata_R <- subset(data.frame(sample_data(ps.rel6)), select = -c(Antibiotic, SampleID, Week, FisabioID, GABAa1, GABAa5, Ymaze_trialstolearn, OLM, IL1_b, Ammonia, Ambulatory_Counts, Vertcal_Counts, Stereotipic_Counts, Average_velocity,  IL_6, TNF_a, IFN_gamma, IL_15, IL_17, IL_10,  IL_4, TGF_b, CCL2,  CCL5, CCL20, CX3CL1,  CCR2,  CCR5, CX3CR1, Occludine, Learning_index_Rmaze))
metadata_R <- metadata_R %>% mutate(Batch = ifelse(Batch=="T5",1,0))
metadata_R <- metadata_R %>% mutate(Treated_CCL = ifelse(Treated_CCL=="Yes",1,0))
metadata_R <- as.data.frame(metadata_R)

# check correct ordering
all(rownames(metadata_R) == rownames(Meta_Species))
## [1] TRUE
all(order(rownames(metadata_R)) == order(rownames(Meta_Species)))
## [1] TRUE
Meta_Species<- as.data.frame(Meta_Species)
Output1_ctrrifvsctrrif <- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                 metaMat = as.data.frame(metadata_R), nnodes = 4)
                                 
Output2_batch_ctrrifvsctrrif<- MetaDeconfound(featureMat = as.data.frame(Meta_Species),
                                     metaMat = as.data.frame(metadata_R), nnodes = 4, randomVar = list("+ (1|Batch)",
                                                                                        c("Batch")))
View(Output1_ctrrifvsctrrif)
View(Output2_batch_ctrrifvsctrrif)

left <- BuildHeatmap(Output1_ctrrifvsctrrif)
right <- BuildHeatmap(Output2_batch_ctrrifvsctrrif)
plot(left)
plot(right)
saveRDS(Output1_ctrrifvsctrrif, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output1_ctrrifvsctrrif.Rds")
saveRDS(Output2_batch_ctrrifvsctrrif, "/fast/AG_Forslund/Lola/CIPF_2018/Analysis/Rats_HE_CCL4model/output/MetadeconfoundR/Output2_batch_ctrrifvsctrrif.Rds")

```

#Save ctrvsctrrif and ccl4vscll4rif comparisons in excel
```{r MetadeconfoundR heatmaps}
Output2_batch_asRandVar_ctrvsctrrif <- readRDS("path/Output2_batch_asRandVar_ctrvsctrrif.Rds")
Output2_batch_ccl4vsccl4rif <- readRDS("path/Output2_batch_ccl4vsccl4rif.Rds")
#ctr
library(openxlsx)

# Create a new Excel workbook
wb_ctrvsctrrif <- createWorkbook()

# Add the data frames as separate sheets in the workbook
addWorksheet(wb_ctrvsctrrif, "Ps")
writeData(wb_ctrvsctrrif, sheet = "Ps", Output2_batch_asRandVar_ctrvsctrrif$Ps, rowNames = TRUE)

addWorksheet(wb_ctrvsctrrif, "Qs")
writeData(wb_ctrvsctrrif, sheet = "Qs", Output2_batch_asRandVar_ctrvsctrrif$Qs, rowNames = TRUE)

addWorksheet(wb_ctrvsctrrif, "Status")
writeData(wb_ctrvsctrrif, sheet = "Status", Output2_batch_asRandVar_ctrvsctrrif$status, rowNames = TRUE)

addWorksheet(wb_ctrvsctrrif, "Ds")
writeData(wb_ctrvsctrrif, sheet = "Ds", Output2_batch_asRandVar_ctrvsctrrif$Ds,  rowNames = TRUE)

# Save the workbook to a file
saveWorkbook(wb_ctrvsctrrif, "output_ctrvsctrrif.xlsx")

#ccl4
library(openxlsx)

# Create a new Excel workbook
wb_ccl4vsccl4rif <- createWorkbook()

# Add the data frames as separate sheets in the workbook
addWorksheet(wb_ccl4vsccl4rif, "Ps")
writeData(wb_ccl4vsccl4rif, sheet = "Ps", Output2_batch_ccl4vsccl4rif$Ps,  rowNames = TRUE)

addWorksheet(wb_ccl4vsccl4rif, "Qs")
writeData(wb_ccl4vsccl4rif, sheet = "Qs", Output2_batch_ccl4vsccl4rif$Qs,  rowNames = TRUE)

addWorksheet(wb_ccl4vsccl4rif, "Status")
writeData(wb_ccl4vsccl4rif, sheet = "Status", Output2_batch_ccl4vsccl4rif$status,  rowNames = TRUE)

addWorksheet(wb_ccl4vsccl4rif, "Ds")
writeData(wb_ccl4vsccl4rif, sheet = "Ds", Output2_batch_ccl4vsccl4rif$Ds,  rowNames = TRUE)

# Save the workbook to a file
saveWorkbook(wb_ccl4vsccl4rif, "output_ccl4vsccl4rif.xlsx")
```
