==> R/6_PERMANOVA_and_CDR.Rmd <==
---
title: "6_Differential_abundance_analysis"
output: html_document
date: "2023-05-03"
---

Using Ps_0 because we cannot model adonis with the data transformed by limma. About removebatcheffect from the limma package:

    - This function is useful for removing unwanted batch effects, associated with hybridization time or other technical variables, ready for plotting or unsupervised analyses such as PCA, MDS or heatmaps. The design matrix is used to describe comparisons between the samples, for example treatment effects, that should not be removed. The function (in effect) fits a linear model to the data, including both batches and regular treatments, then removes the component due to the batch effects.
    - This function is not intended to be used prior to linear modelling. For linear modelling, it is better to include the batch factors in the linear model.

##Check the homogeneity condition
Check that variance homogeneity assumptions hold (to ensure the reliability of the results):

Note the assumption of similar multivariate spread among the groups ie. analogous to variance homogeneity
```{r homogeneity}
library(vegan)
library(microbiome)
ps_0 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_0.Rds")
pseq.rel <- microbiome::transform(ps_0, "compositional")
otu <- abundances(pseq.rel)
dist <- vegdist(t(otu))
meta <- meta(pseq.rel)
#Check homogeneity of variances
anova(betadisper(dist, meta$Group))


pseq.rel <- microbiome::transform(ps, "identity")
otu <- abundances(pseq.rel)
dist <- vegdist(t(otu))
meta <- meta(pseq.rel)
#Check homogeneity of variances
anova(betadisper(dist, meta$Group))

```

## Permanova/ADONIS
```{r PERMANOVA/ADONIS}
library(vegan)
#PERMANOVA/ADONIS
ps_0 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_0.Rds")
pseq.rel <- microbiome::transform(ps_0, "compositional")
otu <- abundances(pseq.rel)
meta <- meta(pseq.rel)
permanova <- adonis(t(otu) ~   Antibiotic + Treated_CCL, strata = meta$Batch,
                    data = meta, permutations=999, method = "bray", by = "margin")

# P-value
permanova$aov.tab

#Using the same data as for Beta Div representation
ps<- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps.Rds")
otu <- abundances(ps)
meta <- meta(ps)
#Antibiotic and Tratment separately
permanova <- adonis(t(otu) ~   Antibiotic + Treated_CCL,
                    data = meta, permutations=999, method = "bray", by = "margin")

# P-value
permanova$aov.tab

#Group (4) for the beta div plot
permanova <- adonis(t(otu) ~   Group,
                    data = meta, permutations=999, method = "bray", by = "margin")

# P-value
permanova$aov.tab
```

```{r Adonis}
metadata <- meta(ps_0)
BC_dist<- phyloseq::distance(ps_0, method="bray", weighted=T)
ordination<- ordinate(ps_0,
                      method="PCoA", distance= BC_dist)
#Adonis2
vegan::adonis(BC_dist~ Batch+ Antibiotic + Treated_CCL , permutations = 999, data = metadata, na.action = F)
```

#Cirrhosis/dysbiosis ratio
```{r Cirrhosis/dysbiosis ratio}
ps_0 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_0.Rds")
library(phyloseq)
library(microbiome)
library(tidyverse)
library(ggpubr)
library(ggsignif)

# Assuming you already have the phyloseq object pseq.rel

# Define the list of taxa families and order you want to keep for different groups
taxa_to_keep_LRC <- c("Lachnospiraceae", "Ruminococcaceae", "Clostridia UCG-014")
#Changed Bacteroidales instead of Bacteroidaceae
taxa_to_keep_EB <- c("Enterobacteriaceae", "Bacteroidales")

# Define the list of groups
groups <- c("ctr", "ctr+rif", "ccl4", "ccl4+rif")

# Initialize an empty data frame to store CDR values for each sample
CDR_data <- data.frame()

# Loop through each sample and calculate CDR
for (sample_id in sample_names(pseq.rel)) {
  # Subset the original phyloseq object to keep only the taxa for the current sample
  pseq.sample <- subset_samples(pseq.rel, SampleID == sample_id)
  
  # Subset the taxa for the current sample
  pseq.filtered_LRC <- subset_taxa(pseq.sample, Family %in% taxa_to_keep_LRC | Order %in% taxa_to_keep_LRC)
  pseq.filtered_EB <- subset_taxa(pseq.sample, Family %in% taxa_to_keep_EB | Order %in% taxa_to_keep_EB)
  
  # Calculate the sum of Lachnospiraceae, Ruminococcaceae, and Clostridialies XIV
  sum_LR_CXIV <- colSums(as.data.frame(taxa_sums(pseq.filtered_LRC)))
  
  # Calculate the sum of Enterobacteriaceae and Bacteroidaceae
  sum_EB <- colSums(as.data.frame(taxa_sums(pseq.filtered_EB)))
  
  # Calculate CDR for the current sample
  CDR <- sum_LR_CXIV / sum_EB
  
  # Create a data frame with CDR values and sample ID
  CDR_data_sample <- data.frame(SampleID = sample_id, CDR)
  
  # Append the data for the current sample to the overall CDR_data
  CDR_data <- rbind(CDR_data, CDR_data_sample)
}

# Print or manipulate the CDR_data as needed
CDR_data<-as.data.frame(CDR_data)
print(CDR_data)
metadata<-data.frame(sample_data(ps_0))
metadata$CDR<-CDR_data$CDR
sample_data(ps_0)<-metadata
my_comparisons <- list( c("ctr", "ctr+rif"), c("ctr+rif", "ccl4"), c("ccl4", "ccl4+rif"), c("ctr+rif", "ccl4+rif"))
as.data.frame(meta(ps_0)) %>%
    arrange() %>%
    mutate(Group = factor(Group, levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))) %>%
    ggplot( aes(x=Group, y=CDR, color = Group)) + 
    stat_compare_means(label= "p.signif", comparisons = my_comparisons)+ geom_violin() +
    theme(axis.text.x = element_text(size = 16), axis.title.y = element_text(size = 16)) +
    xlab("")
```

##Wilcoxon & Welch test using transformed data
Wilcoxon test is a nonparametric test, the data need not be normally distributed.
Welch t-test asumes normally distributed data.
```{r Wilcox }
hist((abundances(ps)), 113)
#Wilcoxon
set.seed(321)
ps.t<- subset_samples(ps, Group %in% c("ctr", "ccl4+rif")) %>% tax_glom(taxrank = 'Genus')
test.taxa <- taxa(ps.t)
pvalue.wilcoxon <- c()
foldchange <- c()
for (taxa in test.taxa) {
    # Create a new data frame for each taxonomic group
    df <- data.frame(Abundance = abundances(ps.t)[taxa,],
                     Group = meta(ps.t)$Group)
    # Calculate pvalue and effect size (difference beween log means)       
    pvalue.wilcoxon[[taxa]] <- wilcox.test(Abundance ~ Group, data = df)$p.value
}
# Correct p-values for multiple testing
pvalue.wilcoxon.adjusted <- p.adjust(pvalue.wilcoxon)
pvalue.wilcoxon.adjusted<- as.data.frame(pvalue.wilcoxon.adjusted)
rownames(pvalue.wilcoxon.adjusted)<- names(pvalue.wilcoxon)

#Welch
pvalue.welch <- c()
foldchange <- c()
for (taxa in test.taxa) {
    # Create a new data frame for each taxonomic group
    df <- data.frame(Abundance = abundances(ps.t)[taxa,],
                     Group = meta(ps.t)$Group)
    # Calculate pvalue and effect size (difference beween log means)       
    pvalue.welch[[taxa]] <- t.test(Abundance ~ Group, data = df)$p.value
}
# Correct p-values for multiple testing
pvalue.welch.adjusted <- p.adjust(pvalue.welch)

```
write.csv(cbind(as.data.frame(pvalue.wilcoxon.adjusted), as.data.frame(tax_table(ps.t))), "Wilcocxon_ccl4+rifvsctr_genus_log10ARSyNdata")
write.csv(as.data.frame(pvalue.wilcoxon.adjusted), "Wilcocxon_ccl4+rifvsctr_genus")
write.csv(as.data.frame(pvalue.welch.adjusted), "Welch_ccl4+rifvsctr_genus")

#Coin. Bloking Batch.
####Wilcoxon test. Using row data.
```{r Wilcox }
library("coin")
#Wilcoxon
sample_data(ps_0)$Read_depth <- sample_sums(ps_0)
pseq.rel <- microbiome::transform(ps_0, "identity")
otu <- abundances(pseq.rel)
hist(otu, 113)

set.seed(321)
ps.t<- pseq.rel %>% subset_samples(Group %in% c("ctr", "ccl4")) %>% tax_glom(taxrank = 'Genus')
test.taxa <- taxa(ps.t)
pvalue_wilcoxon <- c()
foldchange <- c()
for (taxa in test.taxa) {
   # Create a new data frame for each taxonomic group
    df <- data.frame(Abundance = abundances(ps.t)[taxa,],
                     Group = as.factor(meta(ps.t)$Group), Batch = as.factor(meta(ps.t)$Batch))
    
pvalue_wilcoxon[[taxa]] <- coin::wilcox_test(Abundance ~ Group|Batch, data = df)
}
#For getting an object with the adjusted pvalues 
pvalues_wilcoxcoin <- data.frame()
for(i in 1:length(pvalue_wilcoxon))
{
     Lola<-data.frame()
     Lola<-as.data.frame(names(pvalue_wilcoxon[i]))
     Lola[,2]<-p.adjust(pvalue(pvalue_wilcoxon[[i]]))
     pvalues_wilcoxcoin<-rbind(pvalues_wilcoxcoin, Lola)
}

rownames(pvalues_wilcoxcoin)<- pvalues_wilcoxcoin$`names(pvalue_wilcoxon[i])`
write.csv(cbind(as.data.frame(pvalues_wilcoxcoin), as.data.frame(tax_table(ps.t))), "Wilcoxcoin_ctrvsccl4_genus")
```
write.csv(as.data.frame(pvalues_wilcoxcoin), "Wilcoxcoin_ccl4+rifvsctr_genus")
write.csv(as.data.frame(pvalues_wilcoxcoin), "Wilcoxcoin_ccl4vsctr_genus")
write.csv(as.data.frame(pvalues_wilcoxcoin), "Wilcoxcoin_ccl4vsccl4+rif_genus")
write.csv(as.data.frame(pvalues_wilcoxcoin), "Wilcoxcoin_ccl4+rifvsctr")
write.csv(as.data.frame(pvalues_wilcoxcoin), "Wilcoxcoin_ccl4vsctr")
write.csv(as.data.frame(pvalues_wilcoxcoin), "Wilcoxcoin_ccl4vsccl4+rif")

pvalues_wilcoxcoin <- as.data.frame(sapply(pvalue_wilcoxon, pvalue))
pvalues_kruskalcoin <- as.data.frame(sapply(pvalue_kruskal, pvalue))

pvalues_wilcoxcoin <- data.frame()
for(i in 1:length(test.taxa)){
    print(paste(names(pvalue_wilcoxon[i]), p.adjust(pvalue(pvalue_wilcoxon[[i]]))))
}

####Kuskal-wallis (>2 group variables) test. Using non-transformed data.
```{r Wilcox }
library("coin")
#Kruskal-wallis
pseq.rel1 <- microbiome::transform(ps_0, "compositional")
otu <- abundances(pseq.rel1)
hist(otu, 113)

set.seed(321)
ps.t<- pseq.rel1 %>% tax_glom(taxrank = 'Genus')
test.taxa <- taxa(ps.t)
pvalue_kruskal <- c()
foldchange <- c()
for (taxa in test.taxa) {
   # Create a new data frame for each taxonomic group
    df <- data.frame(Abundance = abundances(ps.t)[taxa,],
                     Group = as.factor(meta(ps.t)$Group), Batch = as.factor(meta(ps.t)$Batch))
pvalue_kruskal[[taxa]] <- coin::kruskal_test(Abundance ~ Group|Batch, data = df)
}

pvalues_kruskalcoin <- data.frame()
for(i in 1:length(pvalue_kruskal))
{
     Lola<-data.frame()
     Lola<-as.data.frame(names(pvalue_kruskal[i]))
     Lola[,2]<-p.adjust(pvalue(pvalue_kruskal[[i]]))
     pvalues_kruskalcoin<-rbind(pvalues_kruskalcoin, Lola)
}
rownames(pvalues_kruskalcoin)<- pvalues_kruskalcoin$`names(pvalue_kruskal[i])`
write.csv(cbind(as.data.frame(pvalues_kruskalcoin), as.data.frame(tax_table(ps.t))), "Kruskall_allgroups_genus")

#Kruskal-wallis
pseq.rel1 <- microbiome::transform(ps_0, "compositional")
otu <- abundances(pseq.rel1)
hist(otu, 113)

set.seed(321)
ps.t<- pseq.rel1
test.taxa <- taxa(ps.t)
pvalue_kruskal <- c()
foldchange <- c()
for (taxa in test.taxa) {
   # Create a new data frame for each taxonomic group
    df <- data.frame(Abundance = abundances(ps.t)[taxa,],
                     Group = as.factor(meta(ps.t)$Group), Batch = as.factor(meta(ps.t)$Batch))
pvalue_kruskal[[taxa]] <- coin::kruskal_test(Abundance ~ Group|Batch, data = df)
}

pvalues_kruskalcoin <- data.frame()
for(i in 1:length(pvalue_kruskal))
{
     Lola<-data.frame()
     Lola<-as.data.frame(names(pvalue_kruskal[i]))
     Lola[,2]<-p.adjust(pvalue(pvalue_kruskal[[i]]))
     pvalues_kruskalcoin<-rbind(pvalues_kruskalcoin, Lola)
}
rownames(pvalues_kruskalcoin)<- pvalues_kruskalcoin$`names(pvalue_kruskal[i])`
write.csv(cbind(as.data.frame(pvalues_kruskalcoin), as.data.frame(tax_table(ps.t))), "Kruskall_allgroups_ASVs")
```

####Wilcoxon test. Using clr-transformed data.
```{r Wilcox }
library("coin")
#Wilcoxon
pseq.rel <- microbiome::transform(ps_0, "clr")
otu <- abundances(pseq.rel)
hist(otu, 113)

set.seed(321)
ps.t<- pseq.rel %>% subset_samples(Group %in% c("ctr", "ccl4")) %>% tax_glom(taxrank = 'Genus')
test.taxa <- taxa(ps.t)
pvalue_wilcoxon <- c()
foldchange <- c()
for (taxa in test.taxa) {
   # Create a new data frame for each taxonomic group
    df <- data.frame(Abundance = abundances(ps.t)[taxa,],
                     Group = as.factor(meta(ps.t)$Group), Batch = as.factor(meta(ps.t)$Batch))
    
pvalue_wilcoxon[[taxa]] <- coin::wilcox_test(Abundance ~ Group|Batch, data = df)
}
#For getting an object with the pvalues without adjusting
pvalues_wilcoxcoin <- data.frame()
for(i in 1:length(pvalue_wilcoxon))
{
     Lola<-data.frame()
     Lola<-as.data.frame(names(pvalue_wilcoxon[i]))
     Lola[,2]<-p.adjust(pvalue(pvalue_wilcoxon[[i]]))
     pvalues_wilcoxcoin<-rbind(pvalues_wilcoxcoin, Lola)
}

rownames(pvalues_wilcoxcoin)<- pvalues_wilcoxcoin$`names(pvalue_wilcoxon[i])`
write.csv(cbind(as.data.frame(pvalues_wilcoxcoin), as.data.frame(tax_table(ps.t))), "Wilcoxcoin_ctrvsccl4_genus_clrtransform")
```

#Glm using mvabund
```{r mvabund}
library(mvabund)
sample_data(ps_0)$Read_depth <- sample_sums(ps_0)
ps.t<- ps_0 %>% microbiome::transform("identity") %>% tax_glom(taxrank = 'Genus')
obj <- as.data.frame(t(otu_table(ps.t)))
obj <- mvabund( obj, row.names=rownames(obj), check.rows=FALSE, check.names=TRUE,
         var.names=NULL, neg=FALSE, na.rm=FALSE )
metadata1<-as.data.frame(meta(ps.t))
metadata1$Group[metadata$Group == 'ctr'] <- '1ccctr'
metadata1$Group[metadata$Group == 'ctr+rif'] <- '2ccctr+rif'
metadata1$Group[metadata$Group == 'ccl4'] <- '3ccl4'
metadata1$Group[metadata$Group == 'ccl4+rif'] <- '4ccl4+rif'
metadata1<-metadata1 %>%
    arrange((SampleID))

metadata0<-subset(as.data.frame(meta(ps_0)), select = c(SampleID, Group, Batch, Read_depth))
metadata0 <-mutate(metadata0, Batch = ifelse(Batch=="T5",1,0))

metadata1<-subset(metadata1, select = c(SampleID, Group, Batch, Read_depth))
metadata1 <-mutate(metadata1, Batch = ifelse(Batch=="T5",1,0))

is.mvabund(obj)
plot.mvabund(obj)
boxplot.mvabund(obj)
meanvar.plot(obj)
set.seed(321)
glm <-manyglm(obj~(metadata0$Group) + (metadata0$Read_depth) + (1|metadata0$Batch), family="negative.binomial")
sumglm <- summary.manyglm(glm, resamp = "montecarlo", p.uni = "adjusted")
View(sumglm[["uni.p"]])
glm2 <-manyglm(obj~(metadata1$Group) + (metadata1$Read_depth) + (1|metadata1$Batch), family="negative.binomial")
sumglm2 <- summary.manyglm(glm2, resamp = "montecarlo", p.uni = "adjusted")
View(sumglm2[["uni.p"]])

write.csv(cbind(as.data.frame(sumglm[["uni.p"]]), as.data.frame(tax_table(ps_0))), "DAA_mvabund_notransforms_genus_vsccl4_BlockBatckandAddSeqDepth")
write.csv(cbind(as.data.frame(sumglm2[["uni.p"]]), as.data.frame(tax_table(ps_0))), "DAA_mvabund_notransforms_genus_BlockBatckandAddSeqDepth")
```

#Linear model using ZicoSeq
```{r ZicoSeq}
library(GUniFrac)
#We take the one before removing the zeros because the data is expected to have lots of zeros
ps_2 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_2.Rds")
comm <- as.matrix(as.data.frame(t(otu_table(ps_2))))
metadata <- as.data.frame(metadata)
rownames(metadata)<-metadata$SampleID
meta.dat <- metadata
meta.dat
ZicoSeq.obj <- ZicoSeq(meta.dat = meta.dat, feature.dat = as.matrix(t(comm)), 
                       grp.name = 'Group', adj.name = 'Batch', feature.dat.type = "count",
                       # Filter to remove rare taxa
                       prev.filter = 0.2, mean.abund.filter = 0,  min.prop = 0, 
                       # Winsorization to replace outliers. I dont think it is the best practice, ask Victor
                       is.winsor = FALSE, outlier.pct = 0.03, winsor.end = 'top',
                       # Posterior sampling 
                       is.post.sample = TRUE, post.sample.no = 25, 
                       # Use the square-root transformation
                       link.func = list(function (x) x^0.5), stats.combine.func = max,
                       # Permutation-based multiple testing correction
                       perm.no = 999,  strata = NULL, 
                       # Reference-based multiple stage normalization
                       ref.pct = 0.5, stage.no = 6, excl.pct = 0.2,
                       # Family-wise error rate control
                       is.fwer = TRUE, verbose = TRUE, return.feature.dat = T)
ZicoSeq.plot(ZicoSeq.obj, meta.dat, pvalue.type = 'p.adj.fdr', cutoff = 0.1, text.size = 10, out.dir = NULL, width = 10, height = 6)

ZicoSeq.obj[["p.adj.fwer"]]
write.csv(as.data.frame(ZicoSeq.obj[["p.adj.fwer"]]), "ZicoSeq_ASVs_not_removingZeros")
```

```{r ZicoSeq without zeros}
library(GUniFrac)
#We take the one before removing the zeros because the data is expected to have lots of zeros
ps_0 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_0.Rds")
comm <- as.matrix(as.data.frame(t(otu_table(ps_0))))
metadata <- as.data.frame(metadata)
rownames(metadata)<-metadata$SampleID
meta.dat <- metadata
meta.dat
ZicoSeq.obj <- ZicoSeq(meta.dat = meta.dat, feature.dat = as.matrix(t(comm)), 
                       grp.name = 'Group', adj.name = 'Batch', feature.dat.type = "count",
                       # Filter to remove rare taxa
                       prev.filter = 0.2, mean.abund.filter = 0,  min.prop = 0, 
                       # Winsorization to replace outliers. I dont think it is the best practice, ask Victor
                       is.winsor = FALSE, outlier.pct = 0.03, winsor.end = 'top',
                       # Posterior sampling 
                       is.post.sample = TRUE, post.sample.no = 25, 
                       # Use the square-root transformation
                       link.func = list(function (x) x^0.5), stats.combine.func = max,
                       # Permutation-based multiple testing correction
                       perm.no = 999,  strata = NULL, 
                       # Reference-based multiple stage normalization
                       ref.pct = 0.5, stage.no = 6, excl.pct = 0.2,
                       # Family-wise error rate control
                       is.fwer = TRUE, verbose = TRUE, return.feature.dat = T)
#This one is better, I do not get the warning about rarefaction
ZicoSeq.plot(ZicoSeq.obj, meta.dat, pvalue.type = 'p.adj.fdr', cutoff = 0.1, text.size = 10, out.dir = NULL, width = 10, height = 6)

ZicoSeq.obj[["p.adj.fwer"]]
write.csv(cbind(as.data.frame(ZicoSeq.obj[["p.adj.fwer"]]), as.data.frame(tax_table(ps_0))), "ZicoSeq_ASVs_removingZeros")
```

#Linear modelling using LOCOM
```{r LOCOM}
install.packages(c("Rcpp", "RcppArmadillo", "metap", "utils", "psych", "permute", "parallel", "devtools", "BiocManager"))
BiocManager::install("BiocParallel")
devtools::install_github("yijuanhu/LOCOM")
library(LOCOM)

# loading data
data("metadata")
tax_tab<-as.data.frame(tax_table(ps_0))
otu.table <- as.matrix(as.data.frame(t(otu_table(ps_0))))
Y <- ifelse(metadata$Antibiotic == "Yes", 0, 1)
C <- data.matrix(model.matrix(Y ~ metadata$Batch + metadata$Treated_CCL - 1))[, -1]

# running locom
res <- locom(otu.table = otu.table, Y = Y, C = C[, 1], fdr.nominal = 0.1, seed = 1, adjustment = "Sandev", n.cores = 4)

# summarizing results
# ordering the detected OTUs by their p-values. If no OTU is detected, we can still provide a summary table for
# the top (e.g., 10) OTUs by re-defining o = order(res$p.otu)[1:10]
w <- match(res$detected.otu, colnames(res$p.otu))
o <- w[order(res$p.otu[w])]

summary.table <- data.frame(otu.name = colnames(res$p.otu)[o],
                            mean.freq = colMeans(otu.table/rowSums(otu.table))[o],
                            prop.presence = prop.presence[o],
                            p.value = signif(res$p.otu[o], 3),
                            q.value = signif(res$q.otu[o], 3),
                            effect.size = signif(res$effect.size[o], 3),
                            otu.tax = tax_tab[as.numeric(colnames(res$p.otu)[o]) + 1],
                            row.names = NULL)
summary.table
```

#Top factors
```{r top factors}
# samples x species as input
 ps.taxa.sub <- subset_samples(ps_0, Group %in% c("ccl4", "ccl4+rif"))
pseq.rel <- microbiome::transform(ps.taxa.sub, "identity")
otu <- abundances(pseq.rel)
rownames(otu) <- as.character(tax_table(pseq.rel)[, "Genus"])
rownames(otu) <- gsub("g__", "", rownames(otu))
meta <- meta(pseq.rel)

library(vegan)
permanova <- adonis(t(otu) ~ Group,
               data = meta, permutations=99, method = "horn")

# P-value
print(as.data.frame(permanova$aov.tab)["Group", "Pr(>F)"])

# Analysis of Variance Table
dist <- vegdist(t(otu))
anova(betadisper(dist, meta$Group))


coef <- coefficients(permanova)["Group1",]
top.coef <- coef[rev(order(abs(coef)))[1:30]]
par(mar = c(3, 14, 2, 1))
barplot(sort(top.coef), horiz = T, las = 1, main = "Top taxa")
```

#Graphical representation of ASVs abundance differences
```{r Wilcox representez}
library(tibble)
library(forcats)
set.seed(321)
ps.t<- (ps_0)
df <- data.frame(Abundance = t(abundances(ps.t)),
                 Group = meta(ps.t)$Group)

##Reorder variables
data_new<-factor()
data_new$Group <- factor(df$Group,                 
                         levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))
p2 <- df %>%
  mutate(Group = fct_relevel(Group, 
            "ctr", "ctr+rif", "ccl4", "ccl4+rif")) %>% ggplot(aes(x = Group, y = df$Abundance.ASV1)) +
    geom_boxplot() +
    labs(title = "Absolute abundances", subtitle = "ASV1", y = "Abundance")+ theme(plot.title = element_text(size=18), axis.text.x = element_text(size = 14), axis.title = element_text(size = 16))

p2
```


#HeatMaps of significant taxa
###Coin
Dif abundance coin genus
```{r Heatmap 1}
library("RColorBrewer")
ps.t<- ps_0 %>% subset_samples(Group %in% c("ccl4+rif", "ccl4"))
matrix_1 <- as.matrix(data.frame(otu_table(ps.t)))
matrix_2 <- subset(t(matrix_1), select = c("ASV483", "ASV114", "ASV208", "ASV115", "ASV134", "ASV583", "ASV127", "ASV226", "ASV19", "ASV252", "ASV274", "ASV866", "ASV367", "ASV719", "ASV6", "ASV462", "ASV109", "ASV359", "ASV298", "ASV690"))

metadata_sub <- data.frame(sample_data(ps_0))
matrix <- t(matrix_2)


my_group <- as.numeric(as.factor(meta(ps.t)$Group))
colSide <- t(brewer.pal(8, "Set1")[my_group])
colMain <- colorRampPalette(brewer.pal(8, "Blues"))(16)
heatmap(matrix, Colv = NA, Rowv = NA, scale="row" , ColSideColors=(colSide), col=colMain, xlab="SampleID", ylab="ASV", main="Dif abundance coin genus ccl4+rif vs ccl4")


heatmap(matrix, Colv = NA, Rowv = NA, scale="row", xlab="SampleID", ylab="ASV", main="Dif abundance coin genus ccl4+rif vs ccl4")
```

Dif abundance coin genus
```{r Heatmap 2}
library("RColorBrewer")
ps.t<- ps_0 %>% subset_samples(Group %in% c("ctr", "ccl4"))
matrix_1 <- as.matrix(data.frame(otu_table(ps.t)))
matrix_2 <- subset(t(matrix_1), select = c("ASV358", "ASV583", "ASV18", "ASV127", "ASV58", "ASV213"))

 metadata_sub <- data.frame(sample_data(ps_0))
 matrix <- t(matrix_2)
 

my_group <- as.numeric(as.factor(meta(ps.t)$Group))
colSide <- t(brewer.pal(8, "Blues")[my_group])
colMain <- colorRampPalette(brewer.pal(2, "Blues"))(16)
heatmap(matrix, Colv = NA, Rowv = NA, scale="row" , ColSideColors=(colSide), col=colMain, xlab="SampleID", ylab="ASV", main="Dif abundance coin genus ctr vs ccl4")
 
 
heatmap(matrix, Colv = NA, Rowv = NA, scale="row", xlab="SampleID", ylab="ASV", main="
        Dif abundance coin genus ctr vs ccl4")
```


###mvabund
Dif abundance mvabund ASVs
```{r Heatmap 3}
library(readxl)
DAA_binomialmodel_mvabund_notransforms_ASVs <- read_excel("Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/Dif_abundance_with_ps/mvabund/DAA_binomialmodel_mvabund_notransforms_ASVs.xlsx", 
    sheet = "ctrvsccl4")
View(DAA_binomialmodel_mvabund_notransforms_ASVs)  

ps.t<- ps_0 %>% subset_samples(Group %in% c("ctr", "ccl4"))
DAA_binomialmodel_mvabund_notransforms_ASVs<-DAA_binomialmodel_mvabund_notransforms_ASVs %>% as.data.frame() 

rownames(DAA_binomialmodel_mvabund_notransforms_ASVs)<-DAA_binomialmodel_mvabund_notransforms_ASVs$Compared_to_ctr
matrix_1 <- as.data.frame(cbind(data.frame(otu_table(ps.t)), DAA_binomialmodel_mvabund_notransforms_ASVs))
matrix_2 <- subset(matrix_1, matrix_1$`metadata1$Group3ccl4`<0.05)
matrix_3 <- as.matrix(subset(matrix_2, select = -c(`metadata1$Group3ccl4`, `metadata$BatchT7`, Species, Genus, Family, Order, Class, Phylum, Kingdom)))

matrix <- matrix_3
 
library("RColorBrewer")
my_group <- as.numeric(as.factor(meta(ps.t)$Group))
colSide <- t(brewer.pal(8, "Set1")[my_group])
colMain <- colorRampPalette(brewer.pal(8, "Blues"))(16)
heatmap(matrix, Colv = NA, Rowv = my_group, scale="row" , ColSideColors=(colSide), col=colMain, xlab="SampleID", ylab="ASV", main="Dif abundance mvabund ASV ctr vs ccl4")
```

Dif abundance mvabund ASVs
```{r Heatmap 4}
library(readxl)
DAA_binomialmodel_mvabund_notransforms_ASVs <- read_excel("Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/Dif_abundance_with_ps/mvabund/DAA_binomialmodel_mvabund_notransforms_ASVs.xlsx", 
    sheet = "ccl4vsccl4+rif")
View(DAA_binomialmodel_mvabund_notransforms_ASVs) 

ps.t<- ps_0 %>% subset_samples(Group %in% c("ccl4+rif", "ccl4"))
DAA_binomialmodel_mvabund_notransforms_ASVs<-DAA_binomialmodel_mvabund_notransforms_ASVs %>% as.data.frame() 

rownames(DAA_binomialmodel_mvabund_notransforms_ASVs)<-DAA_binomialmodel_mvabund_notransforms_ASVs$Compared_to_ctr
matrix_1 <- as.data.frame(cbind(data.frame(otu_table(ps.t)), DAA_binomialmodel_mvabund_notransforms_ASVs))
matrix_2 <- subset(matrix_1, matrix_1$`metadata$Groupccl4+rif`<0.05)
matrix_3 <- as.matrix(subset(matrix_2, select = -c(`metadata$Groupccl4+rif`, `metadata$BatchT7`, Species, Genus, Family, Order, Class, Phylum, Kingdom, `Compared_to_ccl4`)))

matrix <- matrix_3
 
library("RColorBrewer")
my_group <- as.numeric(as.factor(meta(ps.t)$Group))
colSide <- t(brewer.pal(8, "Set1")[my_group])
colMain <- colorRampPalette(brewer.pal(8, "Blues"))(16)
heatmap(matrix, Colv = NA, Rowv = my_group, scale="row" , ColSideColors=(colSide), col=colMain, xlab="SampleID", ylab="ASV", main="Dif abundance mvabund ASV ccl4+rif vs ccl4")
```
