==> R/7_correlations_with_metadata.Rmd <==
---
title: "7_correlations_with_metadata"
output: html_document
date: "2023-05-03"
---

#T-tests
```{r Check if there are actual differences}
newmetadata <- metadata %>%
    group_by(Antibiotic) %>%
    filter(!any(Antibiotic == "Yes"))
    
t.test(newmetadata$Average_velocity~newmetadata$Treated_CCL)
t.test(newmetadata$Stereotipic_Counts~newmetadata$Treated_CCL)
t.test(newmetadata$Vertcal_Counts~newmetadata$Treated_CCL)
t.test(newmetadata$Ambulatory_Counts~newmetadata$Treated_CCL)
t.test(newmetadata$GLT1~newmetadata$Treated_CCL)
t.test(newmetadata$GLAST~newmetadata$Treated_CCL)
t.test(newmetadata$OLM~newmetadata$Treated_CCL)
t.test(newmetadata$Learning_index_Rmaze~newmetadata$Treated_CCL)
t.test(newmetadata$Ymaze_trialstolearn~newmetadata$Treated_CCL)

t.test(metadata$Average_velocity~metadata$Treated_CCL)
t.test(metadata$Stereotipic_Counts~metadata$Treated_CCL)
t.test(metadata$Vertcal_Counts~metadata$Treated_CCL)
t.test(metadata$Ambulatory_Counts~metadata$Treated_CCL)
t.test(metadata$GLT1~metadata$Treated_CCL)
t.test(metadata$GLAST~metadata$Treated_CCL)
t.test(metadata$OLM~metadata$Treated_CCL)
t.test(metadata$Learning_index_Rmaze~metadata$Treated_CCL)
t.test(metadata$Ymaze_trialstolearn~metadata$Treated_CCL)

newmetadata2 <- metadata %>%
    group_by(Treated_CCL) %>%
    filter(!any(Treated_CCL == "No"))

t.test(newmetadata2$Average_velocity~newmetadata2$Antibiotic)
t.test(newmetadata2$Stereotipic_Counts~newmetadata2$Antibiotic)
t.test(newmetadata2$Vertcal_Counts~newmetadata2$Antibiotic)
t.test(newmetadata2$Ambulatory_Counts~newmetadata2$Antibiotic)
t.test(newmetadata2$GLT1~newmetadata2$Antibiotic)
t.test(newmetadata2$GLAST~newmetadata2$Antibiotic)
t.test(newmetadata2$OLM~newmetadata2$Antibiotic)
t.test(newmetadata2$Learning_index_Rmaze~newmetadata2$Antibiotic)
t.test(newmetadata2$Ymaze_trialstolearn~newmetadata2$Antibiotic)

t.test(metadata$Average_velocity~metadata$Antibiotic)
t.test(metadata$Stereotipic_Counts~metadata$Antibiotic)
t.test(metadata$Vertcal_Counts~metadata$Antibiotic)
t.test(metadata$Ambulatory_Counts~metadata$Antibiotic)
t.test(metadata$GLT1~metadata$Antibiotic)
t.test(metadata$GLAST~metadata$Antibiotic)
t.test(metadata$OLM~metadata$Antibiotic)
t.test(metadata$Learning_index_Rmaze~metadata$Antibiotic)
t.test(metadata$Ymaze_trialstolearn~metadata$Antibiotic)

kruskal.test(metadata$Average_velocity~metadata$Group)
kruskal.test(metadata$Stereotipic_Counts~metadata$Group)
kruskal.test(metadata$Vertcal_Counts~metadata$Group)
kruskal.test(metadata$Ambulatory_Counts~metadata$Group)
kruskal.test(metadata$GLT1~metadata$Group)
kruskal.test(metadata$GLAST~metadata$Group)
kruskal.test(metadata$OLM~metadata$Group)
kruskal.test(metadata$Learning_index_Rmaze~metadata$Group)
kruskal.test(metadata$Ymaze_trialstolearn~metadata$Group)

```


#Correlations
```{r Heatmap 1}
library(microbiome) # Load libraries
library(phyloseq)
library(dplyr)
library(reshape2)
library(knitr)

#If you wanted to select taxa and group
pseq2 <- ps_0 %>%
         subset_taxa(Phylum == "Bacteroidetes") %>%
         subset_samples(group == "LGG")


# Z transformed abundance data
pseqz <- microbiome::transform(ps_0, "Z")

# Load example data 
ps_3 <- tax_glom(ps_0, taxrank = 'Genus', NArm = FALSE) 
otu <- as.matrix(otu_table(pseqz))

# Define data sets to cross-correlate
metadata1 <- subset(metadata, select = -c(SampleID, Group, Antibiotic, Treated_CCL, FisabioID, Week, Batch))
metadata1 <- as.matrix(metadata1) 
rownames(metadata1) <- (metadata$SampleID)
x <- log10(t(otu))
y <- as.matrix(metadata1) 

# Cross correlate data sets
correlations <- associate(x, y, method = "spearman", mode = "matrix", p.adj.threshold = 0.05, n.signif = 1)

# Or, alternatively, the same output is also available in a handy table format
correlation.table <- associate(x, y, method = "spearman", mode = "table", p.adj.threshold = 0.05, n.signif = 1)

kable(head(correlation.table))
p <- heat(correlation.table, "X1", "X2", 
          fill = "Correlation", 
          star = "p.adj", 
          p.adj.threshold = 0.05) 

p + theme(text=element_text(size=10), 
          axis.text.x = element_text(angle = 90, hjust = 1), 
          legend.key.size = unit(1.3, "cm"))
```

```{r Heatmap 2}
# Z transformed abundance data
pseqz <- microbiome::transform(ps_0, "Z")
otu <- as.matrix(otu_table(pseqz))

# Define data sets to cross-correlate
metadata1 <- subset(metadata, select = -c(SampleID, Group, Antibiotic, Treated_CCL, FisabioID, Week, Batch))
metadata1 <- as.matrix(metadata1) 
rownames(metadata1) <- (metadata$SampleID)
x <- as.data.frame(t(otu))
y <- as.data.frame(metadata1) 
# Set black-and-white theme
library(ggplot2)
theme_set(theme_bw())

# Pick only the correlations with q<0.05
# Note: this will leave other cells empty
library(dplyr)
correlation.table <- associate(x, y, method = "spearman", mode = "table", p.adj.threshold = 0.95, n.signif = 1)
subtable <- filter(correlation.table, p.adj < 0.6)

# Arrange the figure
p <- ggplot(subtable, aes(x = X1, y = X2, fill = Correlation))
p <- p + geom_tile() 
p <- p + scale_fill_gradientn("Correlation", 
                              breaks = seq(from = -1, to = 1, by = 0.2), 
                              colours = c("darkblue", "blue", "white", "red", "darkred"), 
                              limits = c(-1,1)) 

# Polish texts
p <- p + theme(axis.text.x=element_text(angle = 90, hjust=1, face = "italic"),
               axis.text.y=element_text(size = 8))
p <- p + xlab("") + ylab("")

# Mark the most significant cells with stars
p <- p + geom_text(data = subset(correlation.table, p.adj < 0.1), 
                   aes(x = X1, y = X2, label = "+"), col = "white", size = 5)

# Plot
print(p)
```

```{r correlation plot 2}
library(GGally)
ps_3 <- tax_glom(ps_0, taxrank = 'Genus', NArm = FALSE) 
otu <- as.matrix(otu_table(ps_3))
taxa <- as.data.frame(as.matrix(tax_table(ps_3)))


y <- as.data.frame(cbind(taxa, otu)) 
x <- log10(t(otu_table(ps_3)))
colnames(x) <- as.character(tax_table(ps_3)[, "Genus"])

y <- as.matrix(metadata1)

z <- cbind(x, y)
ggcorr(z)
ggcorr(z, label_size = 3, angle = 270, hjust = 0.97)
```

corrplot(cor(z), method = "square",
         title = "method = 'circle'",
         tl.pos = "n", mar = c(2, 1, 3, 1)) 

corrgram(z,
         order = TRUE,              
         upper.panel = panel.pie,   
         lower.panel = panel.shade, 
         text.panel = panel.txt,    
         main = "Correlogram")

metadata2 <- subset(metadata, select = -c(SampleID, Group, Antibiotic, Treated_CCL, FisabioID, Week, Batch, OLM,	GLAST,	GLT1,	GAT1,	GAT3,	TNFR1,	NR2B,	NR2A,	NR1,	GLUA1,	GLUA2,	GABAa1,	GABAa5,	Ambulatory_Counts,	Vertcal_Counts,	Stereotipic_Counts,	Average_velocity,	CCL2,	CCL5,	CCL20,	CX3CL1,	CCR2,	CCR5,	CX3CR1,	Occludine, Ammonia))
Not removed:"IL-6",	"IL1-b",	"TNF-a",	"IFN-gamma",	"IL-15",	"IL-17",	"IL-10",	#"IL-4",	"TGF-b"

#Correlations using coin package
Blocking the group
We need to change the variable that we want to do the correlation... still figuring out how to make a loop so that we don't need to change manually the variable that we want to look at.
```{r Coin spearman}
coin::spearman_test(y$SCFA_AA~x$ASV1|as.factor(metadata$Group))
ps.t <- ps_0
test.taxa <- taxa(ps.t)
spearman_data <- c()
for (taxa in test.taxa) {
   # Create a new data frame for each taxonomic group
    df <- data.frame(Abundance = abundances(ps.t)[taxa,],
                     Group = as.factor(meta(ps.t)$Group),
                     SCFA_AA = as.matrix(meta(ps.t)$SCFA_AA))
    
spearman_data[[taxa]] <- coin::spearman_test(Abundance ~ SCFA_AA|Group, data = df)
}
#For getting an object with the pvalues without adjusting
spearman_ok <- data.frame()
for(i in 1:length(spearman_data))
{
     Lola<-data.frame()
     Lola<-as.data.frame(names(spearman_data[i]))
     Lola[,2]<-p.adjust(pvalue(spearman_data[[i]]))
     Lola[,3]<-(statistic(spearman_data[[i]]))
     colnames(Lola)<- c("Names","pvalue", "zvalue")
     spearman_ok<-rbind(spearman_ok, Lola)
}
View(spearman_ok)
```

#Shared core with venn diagram
Sudarshan Shetty, Leo Lahti, et al.
In may instances one would wish to compare core taxa that are shared between multiple groups. Here, we demonstrate how this can be achieved by microbiome and eulerr. The test data is stored in the microbiomeutilities R package and the original source of data is Zackular et al., 2014: The Gut Microbiome Modulates Colon Tumorigenesis.

devtools::install_github('microsud/microbiomeutilities')

```{r paquetes para diagrama core}
library(eulerr)
library(microbiome)
library(microbiomeutilities)
```

```{r Datos y modificaciones}
pseq5 <- microbiome::transform(ps_0, "identity")
# simple way to count number of samples in each group
table(meta(pseq5)$Group, useNA = "always")

# convert to relative abundances
 pseq.rel5 <- pseq5 %>%
   microbiome::transform(transform = "compositional") %>% tax_glom(taxrank = "Genus")
 Groups5 <- unique(as.character(meta(pseq.rel5 )$Group))

print(Groups5)
```

Write a for loop to go through each of the disease_states one by one and combine identified core taxa into a list.
```{r Imagen Core taxa}
 Groups5 <- unique(as.character(meta(pseq.rel5 )$Group))

list_core <- c() # an empty object to store information

for (n in Groups5){ # for each variable n in Group
    #print(paste0("Identifying Core Taxa for ", n))
    
    ps.sub5 <- subset_samples(pseq.rel5, Groups5 == n) 
    
    core_m <- core_members(ps.sub5, # ps.sub is phyloseq selected with only samples from g 
                           detection = 0.001, # 0.001 in at least 95% samples 
                           prevalence = 0.95)
    print(paste0("No. of core taxa in ", n, " : ", length(core_m))) 
    # print core taxa identified in each Group.
    list_core[[n]] <- core_m # add to a list core taxa for each group.
 }


# Specify colors and plot venn
# supplying colors in the order they appear in list_core
mycols <- c(ctr="#d6e2e9", `ctr+rif` ="#cbf3f0", ccl4="#fcf5c7", `ccl4+rif` = "rosybrown1") 

plot(venn(list_core),
     fills = mycols)
```

Añade: pdf("my_plot.pdf", width = 5.0, height = 6.81) 
pon el script del plot
dev.off()

# Barplots metadata 
```{r Barplots metadata }
ggplot(metadata, aes(x=Group, y=SCFA_AA)) + 
    geom_bar(stat = "identity")


#Reordenado
metadata %>%
    arrange(SCFA_AA) %>%
    mutate(Group = factor(Group, levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))) %>%
    ggplot( aes(x=Group, y=SCFA_AA)) + geom_bar(stat = "identity") +
    theme() +
    xlab("")


#Cambiando los tamaños
metadata %>%
    arrange(SCFA_AA) %>%
    mutate(Group = factor(Group, levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))) %>%
    ggplot(aes(x=Group, y=SCFA_AA)) + geom_bar(stat = "identity") +
    theme(
        axis.text.x = element_text(size = 16), axis.title.y = element_text(size = 16)) +
    xlab("")  
```

#HeatMaps of significant taxa
```{r Heatmap 1}
library("RColorBrewer")
ps.t<- ps_0 %>% subset_samples(Group %in% c("ccl4+rif", "ccl4"))
matrix_1 <- as.matrix(data.frame(otu_table(ps.t)))
matrix_2 <- subset(t(matrix_1), select = c("ASV483", "ASV114", "ASV359", "ASV208", "ASV115", "ASV134", "ASV583", "ASV127", "ASV226", "ASV19", "ASV252", "ASV274", "ASV866", "ASV298", "ASV54", "ASV367", "ASV690", "ASV719", "ASV6", "ASV462", "ASV109"))

 metadata_sub <- data.frame(sample_data(ps_0))
 matrix <- t(matrix_2)
 

my_group <- as.numeric(as.factor(meta(ps.t)$Group))
colSide <- t(brewer.pal(8, "Set1")[my_group])
colMain <- colorRampPalette(brewer.pal(8, "Blues"))(16)
heatmap(matrix, Colv = NA, Rowv = NA, scale="row" , ColSideColors=(colSide), col=colMain    )
 
 
heatmap(matrix, Colv = NA, Rowv = NA, scale="row")
```

```{r Heatmap 2}
library("RColorBrewer")
ps.t<- ps_0 %>% subset_samples(Group %in% c("ctr", "ccl4"))
matrix_1 <- as.matrix(data.frame(otu_table(ps.t)))
matrix_2 <- subset(t(matrix_1), select = c("ASV358", "ASV583", "ASV18", "ASV127", "ASV58", "ASV213"))

 metadata_sub <- data.frame(sample_data(ps_0))
 matrix <- t(matrix_2)
 

my_group <- as.numeric(as.factor(meta(ps.t)$Group))
colSide <- t(brewer.pal(8, "Blues")[my_group])
colMain <- colorRampPalette(brewer.pal(2, "Blues"))(16)
heatmap(matrix, Colv = NA, Rowv = NA, scale="row" , ColSideColors=(colSide), col=colMain    )
 
 
heatmap(matrix, Colv = NA, Rowv = NA, scale="row")
```
