==> R/7_DescriptiveAnalysis_and_CorrelationsMetadata.Rmd <==
---
title: "7_DescriptiveAnalysis_and_CorrelationPlots"
output: html_document
date: "2023-05-03"
---

#Number of taxa per group
```{r descriptive analysis}
ps_0 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_1.Rds")
ps_0
ps.ctr<-ps_0 %>% subset_samples(Group %in% c("ctr"))
ps.ctr<-prune_taxa(taxa_sums(ps.ctr) > 0, ps.ctr)
ps.ctr
ps.ctrrif<-ps_0 %>% subset_samples(Group %in% c("ctr+rif"))
ps.ctrrif<-prune_taxa(taxa_sums(ps.ctrrif) > 0, ps.ctrrif)
ps.ctrrif
ps.ccl4<-ps_0 %>% subset_samples(Group %in% c("ccl4"))
ps.ccl4<-prune_taxa(taxa_sums(ps.ccl4) > 0, ps.ccl4)
ps.ccl4
ps.ccl4rif<-ps_0 %>% subset_samples(Group %in% c("ccl4+rif"))
ps.ccl4rif<-prune_taxa(taxa_sums(ps.ccl4rif) > 0, ps.ccl4rif)
ps.ccl4rif
```

#Dominant taxa
```{r descriptive analysis 2}
library(repmod)
#Phylum level
ps_0 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_0.Rds")
ps.rel<-ps_0 %>% microbiome::transform("compositional") %>% tax_glom(taxrank = 'Phylum', NArm = FALSE) 
# generate a vector containing the full taxonomy path for all OTUs
wholetax <- do.call(paste, c(as.data.frame(tax_table(ps.rel))
                             [c("Kingdom", "Phylum")], 
                             sep = "__"))  # to distinguish from "_" within tax ranks

# turn the otu_table into a data.frame
otu_export <- as.data.frame(t(otu_table(ps.rel)))
tmp <- names(otu_export)

# paste wholetax and OTU_ids together
for(i in 1:length(tmp)){
    names(tmp)[i] = paste(wholetax[i], tmp[i], sep = "__")
}
# overwrite old names
names(otu_export) <- names(tmp)
otu_export<-t(otu_export)
View(as.data.frame(rowSums(otu_export)))
otu_export1<-as.data.frame(rowSums(otu_export))
make_csv_table(otu_export1, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/Dominant_phyla", info = "Dominant_phyla")

#Family level
ps_0 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_0.Rds")
ps.rel2<-ps_0 %>% microbiome::transform("compositional") %>% tax_glom(taxrank = 'Family', NArm = FALSE)
# generate a vector containing the full taxonomy path for all OTUs
wholetax <- do.call(paste, c(as.data.frame(tax_table(ps.rel2))
                             [c("Kingdom", "Phylum", "Class", "Order", "Family")], 
                             sep = "__"))  # to distinguish from "_" within tax ranks

# turn the otu_table into a data.frame
otu_export.fam <- as.data.frame(t(otu_table(ps.rel2)))
tmp <- names(otu_export.fam)

# paste wholetax and OTU_ids together
for(i in 1:length(tmp)){
    names(tmp)[i] = paste(wholetax[i], tmp[i], sep = "__")
}
# overwrite old names
names(otu_export.fam) <- names(tmp)
otu_export.fam<-t(otu_export.fam)
View(as.data.frame(rowSums(otu_export.fam)))
otu_export2<-as.data.frame(rowSums(otu_export.fam))
make_csv_table(otu_export2, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/Dominant_families", info = "Dominant_fam")
```


Añade: pdf("my_plot.pdf", width = 5.0, height = 6.81) 
pon el script del plot
dev.off()

# Barplots metadata 
```{r Barplots metadata }
metadata<-data.frame(sample_data(ps_0))
ggplot(metadata, aes(x=Group, y=SCFA_AA)) + 
    geom_bar(stat = "identity")
#Reordenado
metadata %>%
    arrange(SCFA_AA) %>%
    mutate(Group = factor(Group, levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))) %>%
    ggplot( aes(x=Group, y=SCFA_AA)) + geom_bar(stat = "identity") +
    theme() +
    xlab("")
#Cambiando los tamaños
metadata %>%
    arrange(SCFA_AA) %>%
    mutate(Group = factor(Group, levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))) %>%
    ggplot(aes(x=Group, y=SCFA_AA)) + geom_bar(stat = "identity") +
    theme(
        axis.text.x = element_text(size = 16), axis.title.y = element_text(size = 16)) +
    xlab("")  
```

#Boxplots metadata 
```{r Boxplots metadata}
sample_data(ps_0)$Read_depth <- sample_sums(ps_0)
ps.t<- microbiome::transform(ps_0, "identity")
#Sequencing Depth
as.data.frame(meta(ps.t)) %>%
    arrange() %>%
    mutate(Group = factor(Group, levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))) %>%
    ggplot( aes(x=Group, y=Read_depth)) + geom_boxplot() +
    theme(axis.text.x = element_text(size = 16), axis.title.y = element_text(size = 16)) +
    xlab("")
```

#Abundance Barplots
###Phylum
```{r Barplots Phylum}
ps.rel = transform(ps_0, "compositional")
# agglomerate taxa
glom <- tax_glom(ps.rel, taxrank = 'Phylum', NArm = FALSE)
ps.melt <- psmelt(glom)
# change to character for easy-adjusted level
ps.melt$Phylum <- as.character(ps.melt$Phylum)
ps.melt <- ps.melt %>%
    group_by(Group, Phylum) %>%
    mutate(median=median(Abundance))

#to get the same rows together
ps.melt_sum <- ps.melt %>%
    group_by(Sample,Group,Phylum) %>%
    summarise(Abundance=sum(Abundance))%>%
    arrange(Group) %>%
    mutate(Group = factor(Group, levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif")))

ggplot(ps.melt_sum, aes(x = Sample, y = Abundance, fill = Phylum)) + 
    geom_bar(stat = "identity", aes(fill=Phylum)) + 
    labs(x="", y="%") +
    facet_wrap(~Group, scales= "free_x", nrow=1,  labeller = labeller(Group = c(ctr = "Control", "ctr+rif" = "Control+rif", ccl4 = "CCl4", "ccl4+rif" = "CCl4+rif"))) +
    theme_classic() + 
    theme(strip.background = element_blank(), 
          axis.text.x.bottom = element_text(angle = -90))
```

###Order
```{r Barplots Order}
ps.rel = transform(ps_0, "compositional")
# agglomerate taxa
glom <- tax_glom(ps.rel, taxrank = 'Order', NArm = FALSE)
ps.melt <- psmelt(glom)
# change to character for easy-adjusted level
ps.melt$Order <- as.character(ps.melt$Order)
ps.melt <- ps.melt %>%
    group_by(Group, Order) %>%
    mutate(median=median(Abundance))

#to get the same rows together
ps.melt_sum <- ps.melt %>%
    group_by(Sample,Group,Order) %>%
    summarise(Abundance=sum(Abundance))%>%
    arrange(Group) %>%
    mutate(Group = factor(Group, levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif")))

ggplot(ps.melt_sum, aes(x = Sample, y = Abundance, fill = Order)) + 
    geom_bar(stat = "identity", aes(fill=Order)) + 
    labs(x="", y="%") +
    facet_wrap(~Group, scales= "free_x", nrow=1) +
    theme_classic() + 
    theme(strip.background = element_blank(), 
          axis.text.x.bottom = element_text(angle = -90))
```

###Family
```{r Barplots Family}
ps.rel = transform(ps_0, "compositional")
# agglomerate taxa
glom <- tax_glom(ps.rel, taxrank = 'Family', NArm = FALSE)
ps.melt <- psmelt(glom)
# change to character for easy-adjusted level
ps.melt$Family <- as.character(ps.melt$Family)
ps.melt <- ps.melt %>%
    group_by(Group, Family) %>%
    mutate(median=median(Abundance))

#to get the same rows together
ps.melt_sum <- ps.melt %>%
    group_by(Sample,Group,Family) %>%
    summarise(Abundance=sum(Abundance))%>%
    arrange(Group) %>%
    mutate(Group = factor(Group, levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif")))

ggplot(ps.melt_sum, aes(x = Sample, y = Abundance, fill = Family)) + 
    geom_bar(stat = "identity", aes(fill=Family)) + 
    labs(x="", y="%") +
    facet_wrap(~Group, scales= "free_x", nrow=1) +
    theme_classic() + 
    theme(strip.background = element_blank(), 
          axis.text.x.bottom = element_text(angle = -90))
```

#Venn diagram
Sudarshan Shetty, Leo Lahti, et al.

devtools::install_github('microsud/microbiomeutilities')

```{r paquetes para diagrama core}
library(eulerr)
library(microbiome)
library(tidyverse)
library(microbiomeutilities)
library(VennDiagram)
```


```{r Datos y modificaciones}
ps_0 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_0.Rds")
# convert to relative abundances
 pseq.rel5 <- ps_0 %>%
   microbiome::transform(transform = "compositional") %>% tax_glom(taxrank = "Genus")

 Groups5 <- unique(as.character(meta(pseq.rel5 )$Group))

print(Groups5)
get_latest_annotation <- function(phyloseq_obj) {
  tax_table <- phyloseq_obj@tax_table %>%
    as.data.frame() %>%
    rownames_to_column('ASV') %>%
    as_tibble() %>%
    tidyr::gather('Rank', 'Value', -ASV) %>%
    nest(data = -ASV) %>%
    mutate(TaxaID = map_chr(data, function(x) {
      x %>%
        dplyr::filter(!is.na(Value)) %>%
        magrittr::use_series(Value) %>%
        tail(1)
    })) %>%
    mutate(TaxaUp = map_chr(data, function(x) {
      x %>%
        dplyr::filter(!is.na(Value)) %>%
        magrittr::use_series(Value) %>%
        tail(2) %>% head(1)
    })) %>%
    mutate(TaxaID = paste(ASV, TaxaID)) %>%
    dplyr::select(-c(data, TaxaUp))
  
  return(tax_table)
}

##Usage latest_annotations <- get_latest_annotation(phyloseq_obj)
latest_annotations <- get_latest_annotation(pseq.rel5)
```

Write a for loop to go through each of the disease_states one by one and combine identified core taxa into a list.
```{r Core taxa list}
 Groups5 <- unique(as.character(meta(pseq.rel5 )$Group))

list_core <- c() # an empty object to store information

for (n in Groups5){ # for each variable n in Group
    #print(paste0("Identifying Core Taxa for ", n))
    
    ps.sub5 <- subset_samples(pseq.rel5, Groups5 == n) 
    
    core_m <- core_members(ps.sub5, # ps.sub is phyloseq selected with only samples from g 
                           detection = 0.0001, # 0.0001 in at least 95% samples 
                           prevalence = 0.95)
    print(paste0("No. of core taxa in ", n, " : ", length(core_m))) 
    # print core taxa identified in each Group.
    list_core[[n]] <- core_m # add to a list core taxa for each group.
 }


# Not the venn plot used
# supplying colors in the order they appear in list_core
#mycols <- c(ctr="#d6e2e9", `ctr+rif` ="#cbf3f0", ccl4="#fcf5c7", `ccl4+rif` = "rosybrown1") 

#plot(venn(list_core),
#     fills = mycols)

# Create a named vector of ASV names from the latest_annotations tibble
asv_names <- setNames(latest_annotations$TaxaID, latest_annotations$ASV)

# Replace ASV names in the list_core with the corresponding names from asv_names
list_core <- lapply(list_core, function(x) asv_names[x])

# Print the updated list_core
list_core

# Find common elements
common_elements <- Reduce(intersect, list_core)

# Find specific elements for each list
specific_elements <- lapply(list_core, setdiff, common_elements)

# Print the common elements
cat("Common Elements:\n")
print(common_elements)

# Print the elements specific to each list
cat("\nElements Specific to Each List:\n")
print(specific_elements)


```


```{r plots}
#Make the plot
venn.diagram(
    x = list(ctr=list_core$ctr, `ctr+rif` = list_core$`ctr+rif`, ccl4=list_core$ccl4, `ccl4+rif` = list_core$`ccl4+rif`),
    category.names = c("Control" , "Control+rifaximin" , "CCL4", "CCL4+rifaximin"),
    filename = '~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Figures/venn.png',
    # Output features
    output = TRUE ,
    imagetype="png" ,
    height = 9800 , 
    width = 9800 , 
    resolution = 5000,
    compression = "lzw",
    
    # Circles
    lwd = 1,
    lty = 'blank',
    col=c("#440154ff", '#21908dff', '#fde725ff', 'orange'),
    fill = c(("#440154ff"), ('#21908dff'), ('#fde725ff'), ('orange')),
    
     # Numbers
    cex = 0.5,
    fontface = "bold",
    fontfamily = "sans",
    
    # Set names
    cat.cex = 0.3,
     cat.fontface = "bold",
    cat.default.pos = "outer",
    cat.pos = c(0, 0, 0, 0),
    cat.dist = c(0.055, 0.055, 0.085, 0.085),
    cat.fontfamily = "sans",
    cat.col = c("#440154ff", 'turquoise4', 'gold4', 'orange4')
)

#Plot with names
#library(ggvenn)
#library(RColorBrewer)

#x <- list(Control=list_core$ctr, `Control+rifaximin` = list_core$`ctr+rif`, CCL4=list_core$ccl4, `CCL4+rifaximin` = list_core$`ccl4+rif`)

#ggvenn(x, show_elements = T, digits = 5, label_sep = "\n", fill_color =  c(("#440154ff"), ('#fde725ff'), ('orange'), ('#21908dff')), fill_alpha = 0.65, set_name_color = c("#440154ff", 'gold4', 'orange4', 'turquoise4'), set_name_size = 8, text_size = 3, stroke_linetype = 1, stroke_size = 0.009) + theme(plot.tag = element_text(face = "bold"))
```


#Correlations
```{r Heatmap 1}
ps_0 <- readRDS("~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/ps_0.Rds")
library(microbiome) # Load libraries
library(phyloseq)
library(dplyr)
library(reshape2)
library(knitr)

```

###Correlations SCFAs
```{r correlation plot 1}
newmetadata <- metadata %>%
    subset(select=c(SCFA_AA, SCFA_BA, SCFA_PA, SCFA_CA, SCFA_VA, Group)) %>% data.frame()
newmetadata %>% 
    mutate(Group = factor(Group, 
                          levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))) %>% ggpairs(.,
                    title = "SCFAs per group", 
                    mapping = ggplot2::aes(colour=Group), 
                    lower = list(continuous = wrap("smooth", alpha = 0.3, size=0.1), discrete = "points"), 
                    diag = list(discrete="barDiag", continuous = wrap("densityDiag", alpha=0.5 )), 
                    upper = list(combo = wrap("box_no_facet", alpha=0.5),continuous = wrap("cor", size=4))) + 
    theme(panel.grid.major = element_blank())
```

###Correlations some receptors
```{r correlation plot 2}
library(GGally)
newmetadata <- metadata %>%
    subset(select=c(GLAST, GLT1, TNFR1, NR2B, NR2A, NR1, GLUA1, GLUA2, Group)) %>% data.frame()
newmetadata %>% 
    mutate(Group = factor(Group, 
                          levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))) %>% ggpairs(.,
           title = "Receptor correlations per group", 
           mapping = ggplot2::aes(colour=Group), 
           lower = list(continuous = wrap("smooth", alpha = 0.3, size=0.1), discrete = "points", combo = "blank"), 
           diag = list(discrete="barDiag", continuous = wrap("densityDiag", alpha=0.5 )), 
           upper = list(combo = wrap("box_no_facet", alpha=0.5), continuous = wrap("cor", size=4))) + 
    theme(panel.grid.major = element_blank())    # remove gridlines  
```

###Correlations cognition tests
```{r correlation plot 2}
library(GGally)
newmetadata <- metadata %>%
    subset(select=c(Ammonia, Vertcal_Counts, Stereotipic_Counts, Average_velocity, Ymaze_trialstolearn, Group)) %>% data.frame()
newmetadata %>% 
    mutate(Group = factor(Group, 
                          levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))) %>% 
  ggpairs(., title = "Cognitive tests",
          lower = list(continuous = wrap("smooth", alpha = 0.3, size=0.1), discrete = "points", combo = "blank"), 
          diag = list(discrete="barDiag", continuous = wrap("densityDiag", alpha=0.5 )), 
          upper = list(combo = wrap("box_no_facet", alpha=0.5), continuous = wrap("cor", size=4))) + 
    theme(panel.grid.major = element_blank())    # remove gridlines  
```

###Correlations other metabolites and all boxplots
```{r correlation plot 2}
library(GGally)
newmetadata <- metadata %>%
    subset(select=c(GAT1, GAT3, OLM, IL_6, IL1_b, TNF_a, IFN_gamma, IL_15, IL_17, IL_10, IL_4, TGF_b, CCL2, CCL5, CCL20, CX3CL1, CCR2, CCR5, CX3CR1, Occludine, Learning_index_Rmaze)) %>% data.frame()
ggcorr(newmetadata, label = TRUE, label_alpha = TRUE, label_size = 3, angle = 310, hjust = 0.97, low = 'navy', high = 'orange')
ggcorr(newmetadata, label = TRUE, label_alpha = TRUE, label_size = 3, angle = 310, hjust = 0.97)
#all metadata
ggcorr(metadata, label = TRUE, label_alpha = TRUE, label_size = 3, angle = 310, hjust = 0.97)

#correlograms with other function
newmetadata <- metadata %>%
    subset(select=-c(SampleID, Group, Week, Batch, Antibiotic, FisabioID, Treated_CCL)) 
library(corrplot)
library("Hmisc")
cor_5 <- rcorr(as.matrix(newmetadata))
M <- cor_5$r
p_mat <- cor_5$P
corrplot(M, type = "upper", tl.col = "darkblue", tl.srt = 45)

# use "colorRampPallete" to obtain contionus color scales
library(grDevices)
col <- colorRampPalette(c("turquoise4", "white", "chocolate2"))(20)
corrplot(M,  type = "upper", tl.col = "black", tl.srt = 45, col = col)

#Crossing the non-significant ones
corrplot(M, type = "upper", order = "hclust",  tl.col = "darkblue",
         p.mat = p_mat, sig.level = 0.01)


#Boxplots all metadata
analysisSummary <- metadata %>% 
    mutate(Group = factor(Group, 
                          levels=c("ctr", "ctr+rif", "ccl4", "ccl4+rif"))) %>%
    subset(select=-c(SampleID, Week, Batch, Treated_CCL, FisabioID, Antibiotic)) %>%
    gather(key = variable, value = value, -Group) %>%
    group_by(variable, Group) %>% na.omit() 
p <- ggplot(data = analysisSummary, aes(x=Group, y=value)) + 
    geom_boxplot(aes(fill=Group)) + facet_wrap( ~ variable, scales="free")  +
    theme(axis.text.x = element_text(size = 13), axis.title.y = element_text(size = 13), legend.text = element_text(size = 15))
```



```{r Datos y modificaciones}
ps_0 <- microbiome::transform(ps_0, "identity")
# simple way to count number of samples in each group
table(meta(ps_0)$Group, useNA = "always")

# convert to relative abundances
 pseq.rel5 <- ps_0 %>%
   microbiome::transform(transform = "compositional") %>% tax_glom(taxrank = "Genus")
 Groups5 <- unique(as.character(meta(pseq.rel5 )$Group))

print(Groups5)
```

Write a for loop to go through each of the disease_states one by one and combine identified core taxa into a list.
```{r Imagen Core taxa}
 Groups5 <- unique(as.character(meta(pseq.rel5 )$Group))

list_core <- c() # an empty object to store information

for (n in Groups5){ # for each variable n in Group
    #print(paste0("Identifying Core Taxa for ", n))
    
    ps.sub5 <- subset_samples(pseq.rel5, Groups5 == n) 
    
    core_m <- core_members(ps.sub5, # ps.sub is phyloseq selected with only samples from g 
                           detection = 0.001, # 0.001 in at least 95% samples 
                           prevalence = 0.95)
    print(paste0("No. of core taxa in ", n, " : ", length(core_m))) 
    # print core taxa identified in each Group.
    list_core[[n]] <- core_m # add to a list core taxa for each group.
 }


# Specify colors and plot venn
# supplying colors in the order they appear in list_core
mycols <- c(ctr="#d6e2e9", `ctr+rif` ="#cbf3f0", ccl4="#fcf5c7", `ccl4+rif` = "rosybrown1") 

plot(venn(list_core),
     fills = mycols)
```


#T-tests
```{r Check if there are actual differences Antibiotic all}
library(repmod)
library(tidyverse)
library(readxl)
metadata <- read_excel("Documents/Lola/CIPF_2018/Rats_HE_CCL4model/metadata.xlsx")
View(metadata)

metadata<-data.frame(metadata)

isVariableParametric <- function(data, variable) {
  # Perform Shapiro-Wilk test for normality
  normality_test <- shapiro.test(data[[variable]])
  
  # Perform Levene's test for homoscedasticity
  homoscedasticity_test <- car::leveneTest(data[[variable]], data$Antibiotic)
  
  # Check the p-values of both tests
  normality_p_value <- normality_test$p.value
  homoscedasticity_p_value <- homoscedasticity_test$Pr[1]
  
  # Check if the variable is parametric based on the test results
  is_parametric <- normality_p_value > 0.05 && homoscedasticity_p_value > 0.05
  
  return(is_parametric)
}

analysisSummary <- metadata %>%
    subset(select=-c(SampleID, Group, Week, Batch, Treated_CCL, FisabioID)) %>% 
    gather(key = variable, value = value, -Antibiotic) %>%
    group_by(variable, Antibiotic) %>% na.omit() %>%
    summarise(value=list(value)) %>%
    spread(Antibiotic, value) %>% 
    group_by(variable)

analysisSummary <- analysisSummary %>%
    mutate(parametric = isVariableParametric(metadata, variable))

analysisSummary <-analysisSummary %>% mutate(
    # Save the name of the test for convenient reference later
    test = case_when(
        parametric ~ "Welch's t test", TRUE ~ "Wilcoxon test"
    ),
    # Run the t-test for parametric variables and Wilcoxon signed rank test for non-parametric ones, save the p-value
    p_value = case_when(
        parametric ~ t.test(unlist(Yes), unlist(No))$p.value,
            TRUE ~ wilcox.test(unlist(Yes), unlist(No), paired=FALSE, exact=FALSE)$p.value
    ),
    # Run the test again, but now save the effect size
    statistic = case_when(
        parametric~ t.test(unlist(Yes), unlist(No))$statistic,
            TRUE ~ wilcox.test(unlist(Yes), unlist(No), paired=FALSE, exact=FALSE)$statistic
    ),
)

analysisSummary_ok<-analysisSummary %>% data.frame() %>%
    subset(select=-c(Yes, No))
  
make_csv_table(analysisSummary_ok, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/comparisons_antibiotic", info = "comparisons_antibiotic")
```

```{r Check if there are actual differences Treated_CCL all}
analysisSummary_T <- metadata %>%
    subset(select=-c(SampleID, Group, Week, Batch, Antibiotic, FisabioID)) %>% 
    gather(key = variable, value = value, -Treated_CCL) %>%
    group_by(variable, Treated_CCL) %>% na.omit() %>%
    summarise(value=list(value)) %>%
    spread(Treated_CCL, value) %>% 
    group_by(variable)

analysisSummary_T <- analysisSummary_T %>%
    mutate(parametric = isVariableParametric(metadata, variable))

analysisSummary_T <-analysisSummary_T %>% mutate(
    # Save the name of the test for convenient reference later
    test = case_when(
        parametric ~ "Welch's t test", TRUE ~ "Wilcoxon test"
    ),
    # Run the t-test for parametric variables and Wilcoxon signed rank test for non-parametric ones, save the p-value
    p_value = case_when(
        parametric ~ t.test(unlist(Yes), unlist(No))$p.value,
            TRUE ~ wilcox.test(unlist(Yes), unlist(No), paired=FALSE, exact=FALSE)$p.value
    ),
    # Run the test again, but now save the effect size
    statistic = case_when(
        parametric~ t.test(unlist(Yes), unlist(No))$statistic,
            TRUE ~ wilcox.test(unlist(Yes), unlist(No), paired=FALSE, exact=FALSE)$statistic
    ),
)

analysisSummary_Tok<-analysisSummary_T %>% data.frame() %>%
    subset(select=-c(Yes, No))
make_csv_table(analysisSummary_Tok, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/comparisons_Treated_CCL", info = "comparisons_antibiotic")
```

```{r Check if there are actual differences treated ccl vs ccl4+rif}
newmetadata2 <- metadata %>%
    group_by(Treated_CCL) %>%
    filter(!any(Treated_CCL == "No"))

analysisSummary_A_noccl <- newmetadata2 %>% 
    subset(select=-c(SampleID, Group, Week, Batch, Treated_CCL, FisabioID, GABAa5)) %>%
    gather(key = variable, value = value, -Antibiotic) %>%
    group_by(variable, Antibiotic) %>% na.omit() %>%
    summarise(value=list(value)) %>%
    spread(Antibiotic, value) %>% 
    group_by(variable)

analysisSummary_A_noccl <- analysisSummary_A_noccl %>%
    mutate(parametric = isVariableParametric(metadata, variable))

analysisSummary_A_noccl <-analysisSummary_A_noccl %>% mutate(
    # Save the name of the test for convenient reference later
    test = case_when(
        parametric ~ "Welch's t test", TRUE ~ "Wilcoxon test"
    ),
    # Run the t-test for parametric variables and Wilcoxon signed rank test for non-parametric ones, save the p-value
    p_value = case_when(
        parametric ~ t.test(unlist(Yes), unlist(No))$p.value,
            TRUE ~ wilcox.test(unlist(Yes), unlist(No), paired=FALSE, exact=FALSE)$p.value
    ),
    # Run the test again, but now save the effect size
    statistic = case_when(
        parametric~ t.test(unlist(Yes), unlist(No))$statistic,
            TRUE ~ wilcox.test(unlist(Yes), unlist(No), paired=FALSE, exact=FALSE)$statistic
    ),
)

analysisSummary_A_noccl_ok<-analysisSummary_A_noccl %>% data.frame() %>%
    subset(select=-c(Yes, No))
make_csv_table(analysisSummary_A_noccl_ok, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/analysisSummary_A_noccl_ok", info = "analysisSummary_A_noccl_ok")
```

```{r Check if there are actual differences ctr vs treated ccl}
newmetadata <- metadata %>%
    group_by(Antibiotic) %>%
    filter(!any(Antibiotic == "Yes"))%>%
    subset(select=-c(SampleID, Group, Week, Batch, Antibiotic, FisabioID, CCL5)) 

analysisSummary_T_norif <- newmetadata %>% 
    gather(key = variable, value = value, -Treated_CCL) %>%
    group_by(variable, Treated_CCL) %>% na.omit() %>%
    summarise(value=list(value)) %>%
    spread(Treated_CCL, value) %>% 
    group_by(variable)

analysisSummary_T_norif <- analysisSummary_T_norif %>%
    mutate(parametric = isVariableParametric(metadata, variable))

analysisSummary_T_norif <-analysisSummary_T_norif %>% mutate(
    # Save the name of the test for convenient reference later
    test = case_when(
        parametric ~ "Welch's t test", TRUE ~ "Wilcoxon test"
    ),
    # Run the t-test for parametric variables and Wilcoxon signed rank test for non-parametric ones, save the p-value
    p_value = case_when(
        parametric ~ t.test(unlist(Yes), unlist(No))$p.value,
            TRUE ~ wilcox.test(unlist(Yes), unlist(No), paired=FALSE, exact=FALSE)$p.value
    ),
    # Run the test again, but now save the effect size
    statistic = case_when(
        parametric~ t.test(unlist(Yes), unlist(No))$statistic,
            TRUE ~ wilcox.test(unlist(Yes), unlist(No), paired=FALSE, exact=FALSE)$statistic
    ),
)

analysisSummary_T_norif_ok<-analysisSummary_T_norif %>% data.frame() %>%
    subset(select=-c(Yes, No))
make_csv_table(analysisSummary_T_norif_ok, "~/Documents/Lola/CIPF_2018/Rats_HE_CCL4model/output/Tables/comparisons_Treatment_norif", info = "comparisons_Treatment_norif")
```
